% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_cell_to_keyword}
\alias{add_cell_to_keyword}
\title{Add cell name to keywords that support it (see px-specification)}
\usage{
add_cell_to_keyword(keyword, name)
}
\arguments{
\item{keyword}{String, name of keyword}

\item{name}{String, cell name to add to keyword}
}
\value{
String
}
\description{
Add cell name to keywords that support it (see px-specification)
}
