% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{add_excel_sheet}
\alias{add_excel_sheet}
\title{Add a data frame as a sheet to an Excel workbook}
\usage{
add_excel_sheet(wb, df, sheet_name)
}
\arguments{
\item{wb}{An Excel workbook}

\item{df}{A data frame}

\item{sheet_name}{Name of the sheet}
}
\value{
Nothing
}
\description{
Add a data frame as a sheet to an Excel workbook
}
