% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_language_to_keyword}
\alias{add_language_to_keyword}
\title{Add language, other than the main langue, to keyword}
\usage{
add_language_to_keyword(keyword, main_language, language)
}
\arguments{
\item{keyword}{String, name of keyword}

\item{main_language}{Main language of PX-file}

\item{language}{Lange to add to keyword}
}
\value{
String
}
\description{
Add language, other than the main langue, to keyword
}
\keyword{internal}
