% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{add_main_language}
\alias{add_main_language}
\title{Add boolean main_language column}
\usage{
add_main_language(metadata_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}
}
\value{
A data frame
}
\description{
Add boolean main_language column
}
