% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_sub_key_to_keyword}
\alias{add_sub_key_to_keyword}
\title{Add a sub key word (see px-specification for details on sub keys)}
\usage{
add_sub_key_to_keyword(keyword, name)
}
\arguments{
\item{keyword}{String, name of keyword}

\item{name}{String, name of sub key}
}
\value{
String
}
\description{
Add a sub key word (see px-specification for details on sub keys)
}
