% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_totals.R
\name{add_totals_to_data_table_df}
\alias{add_totals_to_data_table_df}
\title{Wrapper around add_totals}
\usage{
add_totals_to_data_table_df(data_table_df, excel_metadata_path, add_totals)
}
\description{
Wrapper around add_totals
}
