% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_totals.R
\name{add_totals_to_px}
\alias{add_totals_to_px}
\title{Add 'totals' to a px object}
\usage{
add_totals_to_px(px, vars)
}
\arguments{
\item{px}{Px object to add total levels to.}

\item{vars}{List of variables to add total levels to.}
}
\value{
A px object
}
\description{
Wrapper around \link{add_totals} to add total levels to a px object.
}
