% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{add_totals_to_source_data}
\alias{add_totals_to_source_data}
\title{Wrapper aroud add_totals() to get values arguments from metadata}
\usage{
add_totals_to_source_data(metadata_path, source_data_path, add_totals)
}
\arguments{
\item{metadata_path}{Path to metadata}

\item{source_data_path}{Path to source data}

\item{add_totals}{List of variables to add totals to.}
}
\value{
Data frame with added totals
}
\description{
Wrapper aroud add_totals() to get values arguments from metadata
}
