% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_classification}
\alias{age_classification}
\title{Age classification}
\format{
A data frame:
\describe{
\item{valuecode}{Value code}
\item{valuetext}{Value text}
\item{10 years classes}{Aggregtation into 10 years classes}
\item{25 years classes}{Aggregtation into 25 years classes}
}
}
\usage{
age_classification
}
\description{
Example data set to create age classification with aggreations form 10 and
25 years classes.
}
\keyword{datasets}
