% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{aggregation_df}
\alias{aggregation_df}
\title{Get data set from aggregation file}
\usage{
aggregation_df(path)
}
\arguments{
\item{path}{to an aggregation file}
}
\value{
A data frame with columns 'valuecode' (character) and a second column
(ordered) named after the aggregation
}
\description{
Get data set from aggregation file
}
\keyword{internal}
