% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{align_data_frames}
\alias{align_data_frames}
\title{Align data frames}
\usage{
align_data_frames(df_a, df_b)
}
\arguments{
\item{df_a}{Data frame to align}

\item{df_b}{Data frame to align to}
}
\value{
A data frame
}
\description{
Align df_a to have the same columns and column types as df_b.
}
