% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{autopen}
\alias{autopen}
\alias{autopen.px}
\title{AUTOPEN}
\usage{
autopen(x, value)

\method{autopen}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current AUTOPEN is returned. If NULL, AUTOPEN is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change AUTOPEN.
}
