% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{break_long_lines}
\alias{break_long_lines}
\title{Split long strings at commas}
\usage{
break_long_lines(str, max_line_length = 256)
}
\arguments{
\item{str}{String}

\item{max_line_length}{Integer longest allowed line length}
}
\value{
A character vector
}
\description{
Long strings are split so they are no longer than 256 characters and end at
a comma.
}
\keyword{internal}
