% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{change_pivot_variables}
\alias{change_pivot_variables}
\title{Change pivot variables}
\usage{
change_pivot_variables(x, value, pivot)
}
\arguments{
\item{x}{A px object}

\item{pivot}{Pivot type (STUB, HEADING, FIGURES)}

\item{variables}{A character vector of variable codes to change to the pivot
type}
}
\value{
A px object
}
\description{
Change pivot variables
}
\keyword{internal}
