% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{charset}
\alias{charset}
\alias{charset.px}
\title{Set CHARSET}
\usage{
charset(x, value)

\method{charset}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{A character string}
}
\description{
Set CHARSET
}
