% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification_file_section_heading_regexp}
\alias{classification_file_section_heading_regexp}
\title{Regexp that matches a classification file section heading}
\usage{
classification_file_section_heading_regexp()
}
\description{
Regexp that matches a classification file section heading
}
\keyword{internal}
