% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{codepage}
\alias{codepage}
\alias{codepage.px}
\title{CODEPAGE}
\usage{
codepage(x, value)

\method{codepage}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current CODEPAGE is returned. If NULL, CODEPAGE is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change CODEPAGE.
}
