% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{confidential}
\alias{confidential}
\alias{confidential.px}
\title{CONFIDENTIAL}
\usage{
confidential(x, value)

\method{confidential}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current CONFIDENTIAL is returned. If NULL, CONFIDENTIAL is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change CONFIDENTIAL.
}
