% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{contents}
\alias{contents}
\alias{contents.px}
\title{CONTENTS}
\usage{
contents(x, value)

\method{contents}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current CONTENTS is returned. If NULL, an error is thrown because CONTENTS cannot be removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change CONTENTS. CONTENTS cannot be removed because it is a mandatory keyword.
}
