% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{contvariable}
\alias{contvariable}
\alias{contvariable.px}
\title{CONTVARIABLE}
\usage{
contvariable(x, value)

\method{contvariable}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string with the name of the variable to
use as CONTVARIABLE. If missing, the current CONTVARIABLE is returned. If
NULL, CONTVARIABLE is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change CONTVARIABLE.. Setting CONTVARIABLE
indexes several variables in table2. Removing CONTVARIABLE removes the
indexing from table2.
}
