% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{convert_df_to_code}
\alias{convert_df_to_code}
\title{Create code to construct data frame}
\usage{
convert_df_to_code(df)
}
\arguments{
\item{value}{A data frame}
}
\value{
A character vector
}
\description{
Convert data frame to the code nessasary to construct it as a tibble.
}
\keyword{internal}
