% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{convert_value_to_code}
\alias{convert_value_to_code}
\title{Create constructing code}
\usage{
convert_value_to_code(value)
}
\arguments{
\item{value}{Vector of values to create constructors for}
}
\value{
A character vector
}
\description{
Creates code that construct input value.
}
\keyword{internal}
