% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{copyright}
\alias{copyright}
\alias{copyright.px}
\title{COPYRIGHT}
\usage{
copyright(x, value)

\method{copyright}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current COPYRIGHT is returned. If NULL, COPYRIGHT is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change COPYRIGHT.
}
