% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_dummy_tibbles}
\alias{create_dummy_tibbles}
\title{Create a tibble with dummy values}
\usage{
create_dummy_tibbles(dummy_value)
}
\arguments{
\item{dummy_value}{Value to set in all columns}
}
\description{
Create a tibble with dummy values
}
\keyword{internal}
