% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{create_temp_source_data}
\alias{create_temp_source_data}
\title{Save Excel 'Data' sheet as a temporary .rds data set that can be used by
other functions in the current R session.}
\usage{
create_temp_source_data(metadata_path)
}
\arguments{
\item{metadata_path}{Path to metadata file}
}
\description{
Save Excel 'Data' sheet as a temporary .rds data set that can be used by
other functions in the current R session.
}
