% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{creation_date}
\alias{creation_date}
\alias{creation_date.px}
\title{CREATION-DATE}
\usage{
creation_date(x, value)

\method{creation_date}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current
CREATION-DATE is returned. If NULL, CREATION-DATE is removed.}
}
\description{
Inspect or change CREATION-DATE.
}
