% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{descriptiondefault}
\alias{descriptiondefault}
\alias{descriptiondefault.px}
\title{DESCRIPTIONDEFAULT}
\usage{
descriptiondefault(x, value)

\method{descriptiondefault}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current DESCRIPTIONDEFAULT is returned. If NULL, DESCRIPTIONDEFAULT is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change DESCRIPTIONDEFAULT.
}
