% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable2.R
\name{domain}
\alias{domain}
\alias{domain.px}
\title{DOMAIN}
\usage{
domain(x, value)

\method{domain}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string or data frame.
\itemize{
\item Use character to set DOMAIN for all languages and variables.
\item Use data frame with columns 'variable-code', 'language' and
'domain' to set DOMAIN for specific variables.
\item If missing, the current DOMAIN is returned.
\item If NULL, DOMAIN is removed for all variables.
}}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change DOMAIN.
}
