% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{download_px_or_parquet_and_return_path}
\alias{download_px_or_parquet_and_return_path}
\title{Download parquet or PX-file and return path}
\usage{
download_px_or_parquet_and_return_path(url)
}
\arguments{
\item{url}{}
}
\value{
Character
}
\description{
Assumes the file is .px, or parquet.
}
\keyword{internal}
