% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{drop_blank_rows}
\alias{drop_blank_rows}
\title{Drop rows with only NA values}
\usage{
drop_blank_rows(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
A data frame
}
\description{
Drop rows with only NA values
}
\keyword{internal}
