% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable2.R
\name{elimination}
\alias{elimination}
\alias{elimination.px}
\title{ELIMINATION}
\usage{
elimination(x, value)

\method{elimination}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string or data frame.
\itemize{
\item Use character to set ELIMINATION for all languages and variables.
\item Use data frame with columns 'variable-code', 'language' and
'elimination' to set ELIMINATION for specific variables.
\item If missing, the current ELIMINATION is returned.
\item If NULL, ELIMINATION is removed for all variables.
}}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change ELIMINATION.
}
