% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{enumerate_lines}
\alias{enumerate_lines}
\title{Add lines numbers}
\usage{
enumerate_lines(lines)
}
\arguments{
\item{lines}{A character vector}
}
\value{
A character vector
}
\description{
Add '1=', '2=', etc. at start of each line.
}
\keyword{internal}
