% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{error_if_agument_is_not_character_string}
\alias{error_if_agument_is_not_character_string}
\title{Throw error if argument is not a character string}
\usage{
error_if_agument_is_not_character_string(name, value)
}
\description{
Throw error if argument is not a character string
}
\keyword{internal}
