% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{extract_section}
\alias{extract_section}
\title{Get section in .vs or .agg file}
\usage{
extract_section(lines, heading, key = NULL)
}
\arguments{
\item{lines}{A character vector}

\item{heading}{Character, the section heading}

\item{key}{Optional. Character, subkey in section}
}
\description{
Get section in .vs or .agg file
}
\keyword{internal}
