% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{figures}
\alias{figures}
\alias{figures.px}
\title{Set FIGURES}
\usage{
figures(x, ...)

\method{figures}{px}(x, variable)
}
\arguments{
\item{x}{A px object}

\item{variable}{A character vector of variable codes to change to FIGURES.}
}
\value{
A px object
}
\description{
Set which variable is used as figures. The previous figures variable is
changed to STUB
}
