% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{format_data_as_px_lines}
\alias{format_data_as_px_lines}
\title{Turn metadata and data cube into px lines}
\usage{
format_data_as_px_lines(metadata_df, data_cube)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}

\item{data_cube}{Data frame}
}
\value{
A character vector
}
\description{
Turn metadata and data cube into px lines
}
