% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{format_data_df}
\alias{format_data_df}
\title{Format df for px format}
\usage{
format_data_df(data_df, figures_variable)
}
\arguments{
\item{data_df}{Data frame with data in tidy format.}

\item{figures_variable}{Character. The name of the figures variable.}
}
\value{
A data frame
}
\description{
Turn all variables, except figures variable, into character and replace NA
with dash.
}
