% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_time_values}
\alias{format_time_values}
\title{Format time values for px file}
\usage{
format_time_values(values)
}
\arguments{
\item{values}{Time values}
}
\value{
A character vector
}
\description{
Format time values for px file
}
