% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_codelists_metadata}
\alias{get_codelists_metadata}
\title{Get codelists metadata}
\usage{
get_codelists_metadata(excel_metadata_path, data_table_df)
}
\arguments{
\item{excel_metadata_path}{Path to Excel workbook with metadata.}

\item{data_table_df}{A data frame with data}
}
\value{
A data frame
}
\description{
Codelists contains the translation between the codes that are used in the
data table and the labels used for each language. It also contains
information about sort order and numeric precision for each value.
}
