% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_codelists_sheet}
\alias{get_codelists_sheet}
\title{Get sheet 'Codelists' from Excel workbook}
\usage{
get_codelists_sheet(path)
}
\arguments{
\item{path}{Path to Excel workbook (xls/xlsx).}
}
\value{
A data frame
}
\description{
Get sheet 'Codelists' from Excel workbook
}
