% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_data_cube}
\alias{get_data_cube}
\title{Create data cube from source and meta data}
\usage{
get_data_cube(metadata_path, source_data_path)
}
\arguments{
\item{metadata_path}{Path to metadata.}

\item{source_data_path}{Path to source data}
}
\description{
The data cube has one column for each value of HEADING and is sorted by
value. There is one row for each combination of values of STUB variables. The
ordering of STUB variables are set in the metadata.
}
