% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{get_data_table}
\alias{get_data_table}
\title{Return data table from px object}
\usage{
get_data_table(x, value, labels, sort)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame. If missing, the current DATA is returned.
If NULL, all data rows are removed.}

\item{labels}{Optional. Logic or character vector. If TRUE, the data table
is returned with VALUES instead of CODES. By default the VALUES of the main
language are returned, use a character language code to return VALUES for a
specific language.}

\item{sort}{Optional. If TRUE, the data table is returned in the sort order
defined by \code{\link[=px_order]{px_order()}}. If FALSE, the data table is returned as is.}
}
\description{
Return data table from px object
}
\keyword{internal}
