% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_data_table_df}
\alias{get_data_table_df}
\title{Get table data as data frame}
\usage{
get_data_table_df(excel_metadata_path, data_table, add_totals)
}
\arguments{
\item{excel_metadata_path}{Path to Excel workbook with metadata.}

\item{data_table}{A data frame or a path to an \code{.rds} file with data source.
If NULL, the data is taken from the sheet 'Data' in the Excel metadata
workbook.}

\item{add_totals}{A list of variables to add a 'total' level to. The value
of the total level is looked up in 'Variables' xx_elimination. The code for
the level is found in 'Codelists'. The total is a sum of the values in the
variables with type = FIGURES in 'Variables'. NAs are ignored when summing.}
}
\value{
a data frame
}
\description{
Get table data as data frame
}
