% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_data_table_sheet}
\alias{get_data_table_sheet}
\title{Get 'Data table' sheet from Excel workbook}
\usage{
get_data_table_sheet(excel_path)
}
\arguments{
\item{excel_path}{Path to an xlsx workbook}
}
\value{
A data frame.
}
\description{
To be backwards compatible the function returns any sheet starting with
'Data'.
}
