% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_default_encoding}
\alias{get_default_encoding}
\title{Default encoding to read and save px file in}
\usage{
get_default_encoding()
}
\value{
Character
}
\description{
Default encoding to read and save px file in
}
\keyword{internal}
