% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_df_from_data_table_argument}
\alias{get_df_from_data_table_argument}
\title{Get data frame from data table argument}
\usage{
get_df_from_data_table_argument(excel_metadata_path, data_table)
}
\arguments{
\item{excel_metadata_path}{Path to Excel workbook with metadata.}

\item{data_table}{A data frame or a path to an \code{.rds} file with data source.
If NULL, the data is taken from the sheet 'Data' in the Excel metadata
workbook.}
}
\value{
Data frame
}
\description{
Get data frame from data table argument
}
