% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_df.R
\name{get_encoding_from_metadata}
\alias{get_encoding_from_metadata}
\title{Get encoding name from metadata}
\usage{
get_encoding_from_metadata(metadata_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}
}
\value{
Character
}
\description{
Get encoding name from metadata
}
\keyword{internal}
