% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_encoding_from_px_file}
\alias{get_encoding_from_px_file}
\title{Get encoding listed in px file}
\usage{
get_encoding_from_px_file(px_path)
}
\arguments{
\item{px_path}{Path to a px file}
}
\value{
Character
}
\description{
Encoding is listed in CODEPAGE.
}
\keyword{internal}
