% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_excel_sheet}
\alias{get_excel_sheet}
\title{Get specific sheet from Excel workbook}
\usage{
get_excel_sheet(sheet, add_automatically = FALSE)
}
\arguments{
\item{sheet}{String. Sheet to read.}

\item{add_automatically}{Logical. If TRUE, return an empty data frame if the
sheet does not exist.}
}
\value{
A data frame.
}
\description{
Get specific sheet from Excel workbook
}
\keyword{internal}
