% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_figures_variable}
\alias{get_figures_variable}
\title{Get the name of figures variable}
\usage{
get_figures_variable(metadata_path)
}
\arguments{
\item{metadata_path}{Path to metadata}
}
\value{
Character
}
\description{
Get the name of figures variable
}
