% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_figures_variable_from_excel}
\alias{get_figures_variable_from_excel}
\title{Get figures variable from Excel workbook}
\usage{
get_figures_variable_from_excel(excel_path)
}
\arguments{
\item{excel_path}{Path to Excel metadata workbook.}
}
\value{
Character, name of figures variable.
}
\description{
Get figures variable from Excel workbook
}
