% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_general_metadata}
\alias{get_general_metadata}
\title{Get metadata from 'General' sheet in Excel Workbook}
\usage{
get_general_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{Path to metadata file}
}
\description{
Get metadata from 'General' sheet in Excel Workbook
}
