% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{get_mandatory_variables}
\alias{get_mandatory_variables}
\title{A list of which variables should be in each sheet}
\usage{
get_mandatory_variables()
}
\description{
A list of which variables should be in each sheet
}
\keyword{internal}
