% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_metadata}
\alias{get_metadata}
\title{Create metadata for header in PX file}
\usage{
get_metadata(metadata_path, source_data_path)
}
\description{
The metadata is generated from an Excel sheet and from the source data.
}
