% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_metadata_df_from_excel}
\alias{get_metadata_df_from_excel}
\title{Get a data frame with metadata}
\usage{
get_metadata_df_from_excel(excel_metadata_path, data_df)
}
\arguments{
\item{excel_metadata_path}{Path to xlsx workbook with metadata.}

\item{data_df}{Data frame with data in tidy format.}
}
\value{
A data frame
}
\description{
The metadata from the 3 sheets in the Excel workbook are combined into one
data frame, in a format that's both easy to read, query and convert to a
px-file.
}
