% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_df.R
\name{get_metadata_df_from_px}
\alias{get_metadata_df_from_px}
\title{Get metadata df from px object}
\usage{
get_metadata_df_from_px(x)
}
\arguments{
\item{x}{A px object}
}
\value{
A data frame
}
\description{
Get metadata df from px object
}
