% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template.R
\name{get_metadata_template_from_data}
\alias{get_metadata_template_from_data}
\title{Create a minimal metadata template}
\usage{
get_metadata_template_from_data(
  data_df,
  stub_variables = NULL,
  heading_variables = NULL,
  figures_variable = NULL
)
}
\arguments{
\item{data_df}{A data frame with data to create metadata for}

\item{stub_variables}{A character vector with stub variables.}

\item{heading_variables}{A character vector with heading variables.}

\item{figures_variable}{Name of figure variable,}
}
\value{
A data frame
}
\description{
Stub, heading, and figure variables can be controlled  with arguments. If not
provided, one variable is choosen as heading, one as figures, and the
remaining as stub.
}
