% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{get_metdata_df_from_px_lines}
\alias{get_metdata_df_from_px_lines}
\title{Get metadata df from px lines}
\usage{
get_metdata_df_from_px_lines(metadata_lines)
}
\arguments{
\item{metadata_lines}{A character vector with the header of a px file.}
}
\value{
A data frame
}
\description{
Get metadata df from px lines
}
