% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{get_px_keywords}
\alias{get_px_keywords}
\title{Get PX keywords}
\usage{
get_px_keywords()
}
\value{
data frame with all PX keywords
}
\description{
The row order reflects the recommended sort order in the px file specification.
}
\keyword{internal}
