% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_df.R
\name{get_px_metadata_regex}
\alias{get_px_metadata_regex}
\title{Regular expression to parse header in PX-file}
\usage{
get_px_metadata_regex()
}
\value{
A character vector
}
\description{
Regular expression to parse header in PX-file
}
\keyword{internal}
