% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{get_pxfile_encoding}
\alias{get_pxfile_encoding}
\title{Get file encoding listed in px-file}
\usage{
get_pxfile_encoding(px_file_path)
}
\arguments{
\item{px_file_path}{Path to px file}
}
\value{
Character
}
\description{
The encoding is listed under CODEPAGE. If CODEPAGE isen't given, utf-8 is
assumed.
}
