% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_section}
\alias{get_section}
\title{Return lines in section}
\usage{
get_section(lines, head_line)
}
\arguments{
\item{lines}{A character vector}

\item{head_line}{A numeric value, line number of section heading}
}
\description{
Return lines in section
}
\keyword{internal}
