% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_source_data}
\alias{get_source_data}
\title{Get source data}
\usage{
get_source_data(source_data_path, metadata_path)
}
\arguments{
\item{source_data_path}{Path to source data}

\item{metadata_path}{Path to metadata.}
}
\value{
Data frame
}
\description{
Get source data
}
