% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_table2_metadata}
\alias{get_table2_metadata}
\title{Get language dependent general table metadata}
\usage{
get_table2_metadata(excel_metadata_path)
}
\arguments{
\item{excel_metadata_path}{Path to xlsx workbook with metadata.}
}
\value{
A data frame
}
\description{
Get language dependent general table metadata
}
