% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_table_metadata}
\alias{get_table_metadata}
\title{Get general table metadata}
\usage{
get_table_metadata(excel_metadata_path)
}
\arguments{
\item{excel_metadata_path}{Path to xlsx workbook with metadata.}
}
\value{
A data frame
}
\description{
General contains metadata relevant for the entire data set, e.g. its name,
subject, area, contact person, etc.
}
