% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_timeval_type_from_values}
\alias{get_timeval_type_from_values}
\title{Get time scale code from values (see TIMEVAL in px-specification)}
\usage{
get_timeval_type_from_values(values)
}
\arguments{
\item{values}{Values form px file}
}
\value{
A character vector
}
\description{
Get time scale code from values (see TIMEVAL in px-specification)
}
\keyword{internal}
