% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_values_from_time_format}
\alias{get_values_from_time_format}
\title{Get time values from formatted string}
\usage{
get_values_from_time_format(str)
}
\arguments{
\item{str}{String with time values}
}
\value{
A character vector
}
\description{
Get time values from formatted string
}
\keyword{internal}
