% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_df.R
\name{get_variable_label}
\alias{get_variable_label}
\title{Get variable code and label}
\usage{
get_variable_label(metadata_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}
}
\value{
A data frame
}
\description{
Get a data frame with variable codes and label in all languages.
If VARIABLECODE is used, the relation is defined there, but otherwise the
variable code is set to be the main languages' label.
}
