% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_variable_long_names}
\alias{get_variable_long_names}
\title{Get variables and long names}
\usage{
get_variable_long_names(metadata_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}
}
\value{
A data frame
}
\description{
Get a data frame with variable names and there long names in all languages.
If VARIABLECODE is used, the relation is defined there, but otherwise the
variable name is set to be the main languages' long name.
}
