% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_variable_types}
\alias{get_variable_types}
\title{Returns a tibble with variable types}
\usage{
get_variable_types(df)
}
\arguments{
\item{df}{Data frame}
}
\description{
Returns the name listed at the top of each column of a tibble when printed
to the console, .e.g. \if{html}{\out{<int>}}, \if{html}{\out{<char>}}, \if{html}{\out{<dbl>}} ...
}
\keyword{internal}
