% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_variables_metadata}
\alias{get_variables_metadata}
\title{Get metadata from 'Variables' sheet in Excel Workbook}
\usage{
get_variables_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{Path to metadata file}
}
\description{
Get metadata from 'Variables' sheet in Excel Workbook
}
