% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greenlanders}
\alias{greenlanders}
\title{Greenlanders}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 4 columns.
}
\usage{
greenlanders
}
\description{
A fictive data set with demographic data for Greenlanders split in two cohorts.
}
\keyword{datasets}
