% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{guess_file_encoding}
\alias{guess_file_encoding}
\title{Guess encoding of file}
\usage{
guess_file_encoding(path)
}
\arguments{
\item{path}{Path to file}
}
\value{
Character
}
\description{
Guess encoding of file
}
\keyword{internal}
