% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{heading}
\alias{heading}
\alias{heading.px}
\title{HEADING}
\usage{
heading(x, variables)

\method{heading}{px}(x, variables)
}
\arguments{
\item{x}{A px object}

\item{variables}{Optional. A character vector of variable names to change to
HEADING. If missing, the current HEADING variables are returned.}
}
\value{
A px object or a character vector
}
\description{
Inspect or change which variables are used as headings. The heading order is
also changed to the order of the input variables.
}
\seealso{
\code{\link{stub}} \code{\link{figures}}
}
