% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_path_extension}
\alias{is_path_extension}
\title{Check if a path has a specific extension (function factory)}
\usage{
is_path_extension(extension)
}
\description{
Check if a path has a specific extension (function factory)
}
