% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_rds_list}
\alias{is_rds_list}
\title{Check if rds object is properly formatted}
\usage{
is_rds_list(lst)
}
\arguments{
\item{lst}{List to check}
}
\value{
Logic
}
\description{
Check if rds object is properly formatted
}
