% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_languages.R
\name{language}
\alias{language}
\alias{language.px}
\title{LANGUAGE}
\usage{
language(x, value)

\method{language}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current LANGUAGE
is returned. If NULL, LANGUAGE is removed.}
}
\description{
Inspect or change LANGUAGE.
}
\details{
If LANGUAGES is defined, changing LANGUAGE will also add is to LANGUAGES.
}
\seealso{
\code{\link{languages}}
}
