% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lst_distinct_and_arrange}
\alias{lst_distinct_and_arrange}
\title{Get a sorted list of distinct values in list}
\usage{
lst_distinct_and_arrange(lst)
}
\arguments{
\item{lst}{List to sort}
}
\value{
List
}
\description{
Get a sorted list of distinct values in list
}
