% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template.R
\name{make_template}
\alias{make_template}
\title{Make template}
\usage{
make_template(data_df, languages = c("en"), out_path = temp_xlsx_file())
}
\arguments{
\item{data_df}{A data frame to create metadata for}

\item{languages}{A vector of languagecodes}

\item{out_path}{Path to save .xlsx file at}
}
\value{
A data frame
}
\description{
Create a minimal metadata template for a data frame. This function will be
deprecated in the future when px object has been implemented (#179).
}
