% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{merge_named_lists}
\alias{merge_named_lists}
\title{Put two named lists together, remove duplicates and sort}
\usage{
merge_named_lists(lst1, lst2)
}
\arguments{
\item{lst1}{List to merge}

\item{lst2}{List to merge}
}
\value{
List
}
\description{
Put two named lists together, remove duplicates and sort
}
