% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{metamake}
\alias{metamake}
\title{Create an Excel metadata workbook from a px-file}
\usage{
metamake(pxfile_path, out_path, overwrite_xlsx = TRUE)
}
\arguments{
\item{pxfile_path}{Path to px file}

\item{out_path}{Path to save xlsx file at}

\item{overwrite_xlsx}{Should existing metadata workbook be overwritten?}
}
\value{
Nothing
}
\description{
Turn a px-file into an Excel metadata workbook. If pxmake() is run on that
workbook it turns back into an equivalent px-file.
}
