% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromake.R
\name{micromake}
\alias{micromake}
\title{Create micro px files}
\usage{
micromake(x, out_dir = NULL, table_keywords = NULL)
}
\arguments{
\item{x}{A px object.}

\item{out_dir}{Directory to save px files in.}

\item{table_keywords}{Optional. A data frame with columns 'variable' and
'description'. The description is added to the table
for each non-HEADING variable. Currently doesn't support multiple languages.}
}
\value{
Nothing
}
\description{
Split one px object into many small px files (micro files), with count of
the variables in it.
}
\details{
The HEADING variables are use in all the micro files, and a file is created
for each non-HEADING variable. The new px files are saved in a directory
specified by \code{out_dir}.
}
