% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_languages.R
\name{modify_languages_in_px}
\alias{modify_languages_in_px}
\title{Change language in px object}
\usage{
modify_languages_in_px(x, new_languages)
}
\arguments{
\item{x}{A px object}

\item{new_languages}{A character vector}
}
\value{
A px object
}
\description{
Changes languages in 3 of the px tables
}
