% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_languages.R
\name{modify_languages_in_table}
\alias{modify_languages_in_table}
\title{Change language in table}
\usage{
modify_languages_in_table(df, new_languages, keep_vars, align_df)
}
\arguments{
\item{df}{A data frame to change languages in}

\item{new_languages}{A character vector}

\item{keep_vars}{A character vector}

\item{align_df}{A data frame to align with}
}
\value{
A data frame
}
\description{
Change language in table
}
\keyword{internal}
