% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{modify_or_add_row}
\alias{modify_or_add_row}
\title{Add or modify value}
\usage{
modify_or_add_row(
  df,
  lookup_column,
  lookup_column_values,
  modify_column,
  new_value
)
}
\arguments{
\item{df}{Data frame}

\item{lookup_column}{Column to look up}

\item{lookup_column_values}{Values to look up}

\item{modify_column}{Column to modify}

\item{new_value}{New value to modify/add to modify_column}
}
\value{
A data frame
}
\description{
Modify a value in a row of a data frame based on its value in another column.
If the value is not found, a new row is added.
}
