% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mutate_all_vars_to_character}
\alias{mutate_all_vars_to_character}
\title{Change all variables to character}
\usage{
mutate_all_vars_to_character(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
A data frame
}
\description{
Change all variables to character
}
\keyword{internal}
