% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{new_classification}
\alias{new_classification}
\title{Create new classification object}
\usage{
new_classification(name, prestext, domain, df)
}
\arguments{
\item{name}{Optional. Name of the classification.}

\item{prestext}{Optional. Presentation text.}

\item{domain}{Optional. Character vector with domain names. Used to link to
PX-file.}

\item{df}{Optional. A data frame with required column 'valuecode' and
optional column 'valuetext', if the codes have texts. Each additional column
represents an aggregation. The column name is the name of the aggregation.}
}
\value{
A classification object
}
\description{
Constructor for internal functions
}
\keyword{internal}
