% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{new_px}
\alias{new_px}
\title{Create new px object}
\usage{
new_px(
  languages,
  table1,
  table2,
  variables1,
  variables2,
  cells1,
  cells2,
  acrosscell,
  data
)
}
\arguments{
\item{languages}{A data frame with language metadata.}

\item{table1}{A data frame with language independent table metadata.}

\item{table2}{A data frame with language dependent table metadata.}

\item{variables1}{A data frame with language independent variable metadata.}

\item{variables2}{A data frame with language dependent variable metadata.}

\item{cells1}{A data frame with language independent cells metadata.}

\item{cells2}{A data frame with language dependent cells metadata.}

\item{acrosscell}{A data frame with metadata that spans multiple cells.}

\item{data}{A data frame with data.}
}
\value{
A px object
}
\description{
px constructor for internal functions
}
