% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_codelists1.R
\name{order}
\alias{order}
\alias{order.px}
\title{ORDER}
\usage{
order(x, value)

\method{order}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the column 'order' and one or
more of the columns: 'variable-code' and 'code'. If value is missing, all
current orders are returned. If NULL, all orders are removed.}
}
\description{
Inspect or change ORDER.
}
