% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{pivot_variables}
\alias{pivot_variables}
\title{Get names of pivot variables}
\usage{
pivot_variables(x, pivot)
}
\arguments{
\item{x}{A px object}

\item{pivot}{A string, either "STUB", "HEADING" or "FIGURES"}
}
\value{
A character vector of variable codes
}
\description{
Get names of pivot variables
}
