% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_gl}
\alias{population_gl}
\title{Population Greenland}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 4 columns.
}
\source{
\url{https://bank.stat.gl/pxweb/en/Greenland/Greenland__BE__BE01__BE0120/BEXSTA.px/}
}
\usage{
population_gl
}
\description{
A subset of the population count data available in Statistic Greenland's
BEESTA table.
}
\keyword{datasets}
