% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_codelists1.R
\name{handle_codelists1}
\alias{handle_codelists1}
\alias{precision}
\alias{precision.px}
\title{Handle precision and precisionX}
\usage{
handle_codelists1(x, value, column)

precision(x, value)

\method{precision}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the column 'precision' and one or
more of the columns: 'variable-code' and 'code'. If value is missing, all
current precisions are returned. If NULL, all precisions are removed.}

\item{type}{A character string, precision or precisionX}
}
\description{
Inspect or change PRECISION.
}
