% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_autopen}
\alias{px_autopen}
\alias{px_autopen.px}
\title{AUTOPEN}
\usage{
px_autopen(x, value)

\method{px_autopen}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current AUTOPEN is returned. If NULL, AUTOPEN is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change AUTOPEN.
}
\examples{
# Set AUTOPEN
x1 <-
   px(population_gl) |>
   px_autopen('YES')

# Print AUTOPEN
px_autopen(x1)

# Remove AUTOPEN
x2 <- px_autopen(x1, NULL)
px_autopen(x2)
}
