% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_axis_version}
\alias{px_axis_version}
\alias{px_axis_version.px}
\title{AXIS-VERSION}
\usage{
px_axis_version(x, value)

\method{px_axis_version}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current AXIS-VERSION is returned. If NULL, AXIS-VERSION is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change AXIS-VERSION.
}
\examples{
# Set AXIS-VERSION
x1 <-
   px(population_gl) |>
   px_axis_version('2010')

# Print AXIS-VERSION
px_axis_version(x1)

# Remove AXIS-VERSION
x2 <- px_axis_version(x1, NULL)
px_axis_version(x2)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A40\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C699\%2C0\%5D}{Statistics Sweden's documentation}
}
