% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_acrosscells.R
\name{px_cellnote}
\alias{px_cellnote}
\alias{px_cellnote.px}
\title{CELLNOTE}
\usage{
px_cellnote(x, value)

\method{px_cellnote}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with columns 'cellnote' and one or more columns with the names of the STUB and HEADING variables. The 'cellnote' column is the cellnote text, and the STUB/HEADING columns control which cells the note applies to. Use '*' if the note applies to all cells for a variable. If 'value' is missing, the current CELLNOTE is returned. If NULL, CELLNOTE is removed.}
}
\value{
A px object or data frame.
}
\description{
Inspect or change CELLNOTE.
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A40\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C460\%2C0\%5D}{Statistics Sweden's documentation}
}
