% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_acrosscell.R
\name{px_cellnotex}
\alias{px_cellnotex}
\alias{px_cellnotex.px}
\title{CELLNOTEX}
\usage{
px_cellnotex(x, value)

\method{px_cellnotex}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with columns 'cellnotex' and one or more columns with the names of the STUB and HEADING variables. The 'cellnotex' column is the cellnotex text, and the STUB/HEADING columns control which cells the note applies to. Use '*' if the note applies to all cells for a variable.If 'value' is missing, the current CELLNOTEX is returned. If NULL, CELLNOTEX is removed.}
}
\value{
A px object, or a data frame.
}
\description{
Inspect or change CELLNOTEX.
}
