% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{px_classification_from_path}
\alias{px_classification_from_path}
\title{px classification from path}
\usage{
px_classification_from_path(vs_path, agg_paths)
}
\arguments{
\item{vs_path}{Optional. Path to a values set (.vs) file.}

\item{agg_paths}{Optional.
\itemize{
\item If NULL, aggregation paths are automatically taken from the
[Aggreg] field in the .vs file.
\item Use a vector of paths to one or more aggregation files (.agg) to
manually choose aggregations.
\item Use character(0) if aggregations from the .vs files should not be
added automatically.
}}
}
\value{
A classification object
}
\description{
Create px classification from .vs and .agg files
}
\keyword{internal}
