% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_creation_date}
\alias{px_creation_date}
\alias{px_creation_date.px}
\title{CREATION-DATE}
\usage{
px_creation_date(x, value)

\method{px_creation_date}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current CREATION-DATE is returned. If NULL, CREATION-DATE is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change CREATION-DATE.
}
\examples{
# Set CREATION-DATE
x1 <-
   px(population_gl) |>
   px_creation_date('19960612 14:20')

# Print CREATION-DATE
px_creation_date(x1)

# Remove CREATION-DATE
x2 <- px_creation_date(x1, NULL)
px_creation_date(x2)
}
