% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{px_data}
\alias{px_data}
\alias{px_data.px}
\title{DATA}
\usage{
px_data(x, value)

\method{px_data}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{A data frame}
}
\value{
A px object or data frame.
}
\description{
Inspect or change DATA.
}
