% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_datasymbolnil}
\alias{px_datasymbolnil}
\alias{px_datasymbolnil.px}
\title{DATASYMBOLNIL}
\usage{
px_datasymbolnil(x, value, validate)

\method{px_datasymbolnil}{px}(x, value, validate = TRUE)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current DATASYMBOLNIL is returned. If NULL, DATASYMBOLNIL is removed.}

\item{validate}{Optional. If TRUE a number of validation checks are performed on the px object, and an error is thrown if the object is not valid. If FALSE, the checks are skipped, which can be usefull for large px objects where the check can be time consuming. Use \code{\link[=px_validate]{px_validate()}} to manually preform the check.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change DATASYMBOLNIL.
}
\examples{
# Set DATASYMBOLNIL for all languages
x1 <-
  px(population_gl) |>
  px_datasymbolnil('missing')

# Print DATASYMBOLNIL
px_datasymbolnil(x1)

# Set DATASYMBOLNIL for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_datasymbolnil(tribble(~language, ~value,
                      'en', 'missing',
                      'kl', 'amigaataapput'))
px_datasymbolnil(x2)

# Remove DATASYMBOLNIL
x3 <- px_datasymbolnil(x2, NULL)
px_datasymbolnil(x3)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A44\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C309\%2C0\%5D}{Statistics Sweden's documentation}
}
