% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_descriptiondefault}
\alias{px_descriptiondefault}
\alias{px_descriptiondefault.px}
\title{DESCRIPTIONDEFAULT}
\usage{
px_descriptiondefault(x, value)

\method{px_descriptiondefault}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current DESCRIPTIONDEFAULT is returned. If NULL, DESCRIPTIONDEFAULT is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change DESCRIPTIONDEFAULT.
}
\examples{
# Set DESCRIPTIONDEFAULT
x1 <-
   px(population_gl) |>
   px_descriptiondefault('YES')

# Print DESCRIPTIONDEFAULT
px_descriptiondefault(x1)

# Remove DESCRIPTIONDEFAULT
x2 <- px_descriptiondefault(x1, NULL)
px_descriptiondefault(x2)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A45\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C440\%2C0\%5D}{Statistics Sweden's documentation}
}
