% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{px_figures}
\alias{px_figures}
\alias{px_figures.px}
\title{Change figures variable}
\usage{
px_figures(x, variable)

\method{px_figures}{px}(x, variable)
}
\arguments{
\item{x}{A px object}

\item{variable}{Optional. Name of variable to use as FIGRUES. If missing, the
current PX_FIGURES variable is returned.}
}
\value{
A px object or a character string
}
\description{
Inspect or change which variable is used as figures. The previous figures
variable is changed to STUB. There can only be one figures variable.
}
\examples{
x1 <- px(population_gl)

# Print FIGURES
px_figures(x1)

# Change 'age' to FIGURES variable, 'n' i changed to STUB
x2 <- px_figures(x1, 'age')
px_figures(x2)
px_stub(x2)
}
\seealso{
\code{\link{px_stub}} \code{\link{px_heading}}
}
