% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_df.R
\name{px_from_data_df}
\alias{px_from_data_df}
\title{Create a minimal px object from a data frame}
\usage{
px_from_data_df(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A px object
}
\description{
Create a minimal px object from a data frame
}
\keyword{internal}
