% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_file.R
\name{px_from_px_file}
\alias{px_from_px_file}
\title{Create a px object form a px-file}
\usage{
px_from_px_file(path)
}
\arguments{
\item{path}{Path to a px-file}
}
\value{
A px object
}
\description{
Create a px object form a px-file
}
\keyword{internal}
