% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_map}
\alias{px_map}
\alias{px_map.px}
\title{MAP}
\usage{
px_map(x, value)

\method{px_map}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current MAP is returned. If NULL, MAP is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change MAP.
}
\examples{
# Set MAP for all languages
x1 <-
  px(population_gl) |>
  px_map('Great Britain')

# Print MAP
px_map(x1)

# Set MAP for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_map(tribble(~language, ~value,
                      'en', 'Great Britain',
                      'kl', 'Kalaallit Nunaat'))
px_map(x2)

# Remove MAP
x3 <- px_map(x2, NULL)
px_map(x3)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A49\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C354\%2C0\%5D}{Statistics Sweden's documentation}
}
