% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_matrix}
\alias{px_matrix}
\alias{px_matrix.px}
\title{MATRIX}
\usage{
px_matrix(x, value)

\method{px_matrix}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current MATRIX is returned. If NULL, an error is thrown because MATRIX cannot be removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change MATRIX.
}
\examples{
# Set MATRIX
x1 <-
   px(population_gl) |>
   px_matrix('POPULATION')

# Print MATRIX
px_matrix(x1)
}
