% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_next_update}
\alias{px_next_update}
\alias{px_next_update.px}
\title{NEXT-UPDATE}
\usage{
px_next_update(x, value)

\method{px_next_update}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current NEXT-UPDATE is returned. If NULL, NEXT-UPDATE is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change NEXT-UPDATE.
}
