% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_cells1.R
\name{px_precision}
\alias{px_precision}
\alias{px_precision.px}
\title{PRECISION}
\usage{
px_precision(x, value)

\method{px_precision}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the columns 'precision' and one or more of the columns: 'variable-code', and 'code'. If 'value' is missing, the current PRECISION is returned. If NULL, PRECISION is removed.}
}
\value{
A px object or a data frame
}
\description{
Inspect or change PRECISION.
}
