% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{px_save}
\alias{px_save}
\title{Save px object to file}
\usage{
px_save(x, path, save_data = TRUE, data_path = NULL)
}
\arguments{
\item{x}{A px object.}

\item{path}{Path to file. The file extension determines the format. Can be:
\itemize{
\item \code{.px} to save as a PX-file
\item \code{.xlsx} to save as an Excel metadata workbook
}}

\item{save_data}{If FALSE, no 'Data' sheet is created in the Excel workbook.
Can only be used if \code{path} is an \code{.xlsx} file.}

\item{data_path}{Path to an \code{.rds} or \code{.parquet} file to save data table at.
This is usefull when saving an Excel workbook where the data has more rows
than Excel can handle. Can only be used if 'path' is an \code{.xlsx} file, and
'save_data' is TRUE.}
}
\value{
Nothing
}
\description{
Save px object to file
}
\details{
Use \code{px_codepage()} to change file encoding.
}
\seealso{
\code{\link[=px_codepage]{px_codepage()}}
}
