% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{px_save_classification}
\alias{px_save_classification}
\title{Save classification as .vs and .agg files}
\usage{
px_save_classification(c, path)
}
\arguments{
\item{c}{A classification object}

\item{path}{Directory to save the files in}
}
\value{
Nothing
}
\description{
Save a classification object as .vs and .agg files. The .vs file contains
the value set and the .agg files contain the aggregations.
}
\examples{
c <- px_classification(name = "Age5",
                       prestext = "Ages 0-9 - 60+",
                       domain = "age",
                       df = age_classification
                       )

px_save_classification(c, path = tempdir())
}
