% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_showdecimals}
\alias{px_showdecimals}
\alias{px_showdecimals.px}
\title{SHOWDECIMALS}
\usage{
px_showdecimals(x, value)

\method{px_showdecimals}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current SHOWDECIMALS is returned. If NULL, SHOWDECIMALS is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change SHOWDECIMALS.
}
\examples{
# Set SHOWDECIMALS
x1 <-
   px(population_gl) |>
   px_showdecimals('2')

# Print SHOWDECIMALS
px_showdecimals(x1)

# Remove SHOWDECIMALS
x2 <- px_showdecimals(x1, NULL)
px_showdecimals(x2)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A53\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C305\%2C0\%5D}{Statistics Sweden's documentation}
}
