% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_stockfa}
\alias{px_stockfa}
\alias{px_stockfa.px}
\title{STOCKFA}
\usage{
px_stockfa(x, value)

\method{px_stockfa}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current STOCKFA is returned. If NULL, STOCKFA is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change STOCKFA.
}
\examples{
# Set STOCKFA for all languages
x1 <-
  px(population_gl) |>
  px_stockfa('S')

# Print STOCKFA
px_stockfa(x1)

# Set STOCKFA for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_stockfa(tribble(~language, ~value,
                      'en', 'S',
                      'kl', 'F'))
px_stockfa(x2)

# Remove STOCKFA
x3 <- px_stockfa(x2, NULL)
px_stockfa(x3)
}
