% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{px_timeval}
\alias{px_timeval}
\alias{px_timeval.px}
\title{TIMEVAL}
\usage{
px_timeval(x, variable)

\method{px_timeval}{px}(x, variable)
}
\arguments{
\item{x}{A px object}

\item{variable}{Optional. Name of variable to use as TIMEVAL. If missing, the
current TIMEVAL variable is returned.}
}
\value{
A px object or a character string
}
\description{
Inspect or change which variable is used as px_timeval. There can only be one
time variable.
}
