% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable2.R
\name{px_variable_label}
\alias{px_variable_label}
\alias{px_variable_label.px}
\title{Change variable label}
\usage{
px_variable_label(x, value)

\method{px_variable_label}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string or data frame.
\itemize{
\item Use character to set variable-label for all languages and variables.
\item Use data frame with columns 'variable-code', 'language' and
'variable-label' to set variable-label for specific variables.
\item If missing, the current variable-label is returned.
\item If NULL, variable-label is removed for all variables.
}}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change variable label.

The variable label is the name that is shown in the px file.
}
