% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{pxsave}
\alias{pxsave}
\title{Save px object to file}
\usage{
pxsave(px, path)
}
\arguments{
\item{px}{A px object.}

\item{path}{Path to file. The file extension determines the format. Can be:
\itemize{
\item \code{.px} to save as a px-file
\item \code{.xlsx} to save as an Excel metadata workbook
}}
}
\value{
Nothing
}
\description{
Save px object to file
}
