% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_px_file.R
\name{read_excel_metadata_sheet}
\alias{read_excel_metadata_sheet}
\title{Read specific sheet in Excel metadata file}
\usage{
read_excel_metadata_sheet(table_name, sheet_name)
}
\description{
Read specific sheet in Excel metadata file
}
