% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_lines_guess_encoding}
\alias{read_lines_guess_encoding}
\title{Read lines from file with guessed encoding}
\usage{
read_lines_guess_encoding(path)
}
\arguments{
\item{path}{Path to file}
}
\value{
A character vector
}
\description{
Wrapper around readLines to guess encoding before reading.
}
\keyword{internal}
