% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_lines_with_encoding}
\alias{read_lines_with_encoding}
\title{Read lines with specific encoding}
\usage{
read_lines_with_encoding(path, encoding)
}
\arguments{
\item{path}{Path to file}

\item{encoding}{Encoding to use}
}
\value{
A character vector
}
\description{
Read lines with specific encoding
}
\keyword{internal}
