% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{replace_na_language_with_main_language}
\alias{replace_na_language_with_main_language}
\title{Impute missing language}
\usage{
replace_na_language_with_main_language(metadata_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}
}
\value{
A data frame
}
\description{
Add main language as language if language is missing.
}
