% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_file.R
\name{save_px_as_px_file}
\alias{save_px_as_px_file}
\title{Save px object to px-file}
\usage{
save_px_as_px_file(x, path)
}
\arguments{
\item{x}{A px object}

\item{path}{Path to save px-file at}
}
\value{
Nothing
}
\description{
Save px object to px-file
}
