% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{save_px_as_xlsx}
\alias{save_px_as_xlsx}
\title{Save px object as an Excel workbook}
\usage{
save_px_as_xlsx(px, path)
}
\arguments{
\item{px}{A px object.}

\item{path}{Path to save Excel workbook.}
}
\value{
Nothing
}
\description{
Save px object as an Excel workbook
}
