% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{save_pxmake_output}
\alias{save_pxmake_output}
\title{Save output form pxmake to out_path}
\usage{
save_pxmake_output(out_path, metadata_df, data_table_df)
}
\arguments{
\item{out_path}{Path to save output at, either as an \code{.rds} or \code{.px} file.
If NULL, no file is saved.}

\item{metadata_df}{Data frame with metadata.}

\item{data_table_df}{Data frame with data table in tidy format.}
}
\value{
rds object
}
\description{
Save output form pxmake to out_path
}
