% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{save_temp_data}
\alias{save_temp_data}
\title{Save data set to temporary file}
\usage{
save_temp_data(df)
}
\arguments{
\item{df}{Data frame to save.}
}
\value{
Path to temporary data
}
\description{
Create a temporary .rds data file that can be used by other functions in the
current R session.
}
