% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{section_interval}
\alias{section_interval}
\title{Get index of first and last line of section}
\usage{
section_interval(lines, head_line)
}
\arguments{
\item{lines}{A character vector}

\item{head_line}{A numeric value, line number of section heading}
}
\description{
Get index of first and last line of section
}
\keyword{internal}
