% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{set_data_factors_based_on_ordering_and_sort}
\alias{set_data_factors_based_on_ordering_and_sort}
\title{Set data table vars as factors}
\usage{
set_data_factors_based_on_ordering_and_sort(x)
}
\description{
Set data table vars as factors
}
\keyword{internal}
