% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{smallest_larger_value}
\alias{smallest_larger_value}
\title{Smallest larger value}
\usage{
smallest_larger_value(vec, value)
}
\arguments{
\item{vec}{A numeric vector}

\item{value}{A numeric value}
}
\value{
The smallest value in 'vec' that is larger than 'value'
}
\description{
Smallest larger value
}
\keyword{internal}
