% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{sort_metadata}
\alias{sort_metadata}
\title{Sort metdata keywords in recommended order}
\usage{
sort_metadata(metadata)
}
\arguments{
\item{metadata}{Data frame with metadata.}
}
\value{
data frame
}
\description{
Sort metdata keywords in recommended order
}
