% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_df.R
\name{sort_metadata_df}
\alias{sort_metadata_df}
\title{Sort metadata data frame}
\usage{
sort_metadata_df(metadata_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}
}
\value{
A data frame
}
\description{
The data frame is first sorted by the keyword order defined in the
px-specification and then by the language order.
}
\keyword{internal}
