% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{stockfa}
\alias{stockfa}
\alias{stockfa.px}
\title{STOCKFA}
\usage{
stockfa(x, value)

\method{stockfa}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current STOCKFA is returned. If NULL, STOCKFA is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change STOCKFA.
}
