% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{str_extract_px_encoding}
\alias{str_extract_px_encoding}
\title{Get encoding from CODEPAGE string}
\usage{
str_extract_px_encoding(str, default = "utf-8")
}
\arguments{
\item{str}{Character containing CODEPAGE and its values.}

\item{default}{Character of default encoding to return in no match is found.}
}
\value{
Character
}
\description{
Get encoding from CODEPAGE string
}
