% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{subject_area}
\alias{subject_area}
\alias{subject_area.px}
\title{SUBJECT-AREA}
\usage{
subject_area(x, value)

\method{subject_area}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current SUBJECT-AREA is returned. If NULL, an error is thrown because SUBJECT-AREA cannot be removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change SUBJECT-AREA. SUBJECT-AREA cannot be removed because it is a mandatory keyword.
}
