% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{tableid}
\alias{tableid}
\alias{tableid.px}
\title{TABLEID}
\usage{
tableid(x, value)

\method{tableid}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current TABLEID is returned. If NULL, TABLEID is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change TABLEID.
}
