% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{temp_dir}
\alias{temp_dir}
\title{Create and return path to temporary directory}
\usage{
temp_dir()
}
\value{
Character
}
\description{
Create and return path to temporary directory
}
\keyword{internal}
