% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{temp_file_with_extension}
\alias{temp_file_with_extension}
\title{Create temporary file}
\usage{
temp_file_with_extension(extension)
}
\arguments{
\item{extension}{String, file name extension}
}
\value{
Path to temporary file
}
\description{
Get a temporary file path with a specific extension (function factory)
}
