% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{time_variable}
\alias{time_variable}
\title{Get name of time variable}
\usage{
time_variable(x)
}
\arguments{
\item{x}{A px object}
}
\value{
Variable code of time variable
}
\description{
Get name of time variable
}
