% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{timeval}
\alias{timeval}
\alias{timeval.px}
\title{Set TIMEVAL}
\usage{
timeval(x, variable)

\method{timeval}{px}(x, variable)
}
\arguments{
\item{x}{A px object}

\item{variable}{Character, variable code to change to TIMEVAL.}
}
\value{
A px object
}
\description{
Set which variable is used as time. There can only be one time variable.
}
