% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{title}
\alias{title}
\alias{title.px}
\title{TITLE}
\usage{
title(x, value)

\method{title}{px}(x, value)
}
\arguments{
\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current CONTENTS is returned. If NULL, TITLE is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change TITLE.
}
