% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{units}
\alias{units}
\alias{units.px}
\title{UNITS}
\usage{
units(x, value)

\method{units}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages,
or a data frame with columns 'language' and 'value' to set it for specific
languages. If 'value' is missing, the current UNITS is returned.}
}
\description{
Inspect or change UNITS. UNITS cannot be removed because it is a mandatory
keyword.
}
