% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{update_data_table}
\alias{update_data_table}
\title{Add value as data table in x}
\usage{
update_data_table(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame. If missing, the current DATA is returned.
If NULL, all data rows are removed.}
}
\description{
Add value as data table in x
}
\keyword{internal}
