% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_languages.R
\name{used_languages}
\alias{used_languages}
\title{Get languages used in px object}
\usage{
used_languages(x)
}
\arguments{
\item{x}{A px object}
}
\value{
A character vector
}
\description{
Get languages used in px object
}
