% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_metamake_arguments}
\alias{validate_metamake_arguments}
\title{Check all metamake arguments}
\usage{
validate_metamake_arguments(input, out_path, data_path, create_data)
}
\arguments{
\item{input}{Input can be provided in one of four ways:
\enumerate{
\item A path to a \code{.px} file.
\item A path to a \code{.rds} file created by \link{pxmake}.
\item A named list with two data frames "metadata" and "data" (same as option 2).
\item A data frame with data. A minimal metadata template will be created.
}}

\item{out_path}{Path to save metadata at. Use \code{.xlsx} extension to save
as an Excel workbook. Use \code{.rds} to save as an rds file. If NULL, no file is
saved.}

\item{data_path}{Path to save data as an .rds file. If NULL, the data is
saved as the sheet 'Data' in the Excel metadata workbook.}

\item{create_data}{Logic. If FALSE the data table is not generated. This can
be used to only generate metadata.}
}
\value{
Nothing
}
\description{
Throw an error if there are problems with the function arguments.
}
