% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_micromake_arguments}
\alias{validate_micromake_arguments}
\title{Check all arguments to micromake()}
\usage{
validate_micromake_arguments(px, out_dir)
}
\arguments{
\item{px}{A px object.}

\item{out_dir}{Directory to save px files in.}
}
\value{
Nothing
}
\description{
Check all arguments to micromake()
}
