% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{validate_px}
\alias{validate_px}
\title{Validate px object}
\usage{
validate_px(x)
}
\arguments{
\item{x}{A supposed px object.}
}
\value{
A valid px object.
}
\description{
Throws an error if the px object is not valid.
}
