% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_px_arguments}
\alias{validate_px_arguments}
\title{Check all arguments to px()}
\usage{
validate_px_arguments(input, data)
}
\arguments{
\item{input}{Path to px-file, path to an Excel metadata workbook, or a list.}

\item{data}{Either a data frame or a path to an \code{.rds} file with a data frame.
This can only be used if the \code{input} argument is an Excel metadata workbook.
If NULL, the data should be provided in the 'data' sheet of the Excel
workbook.}
}
\value{
Nothing
}
\description{
Check all arguments to px()
}
