% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_px_classification_arguments}
\alias{validate_px_classification_arguments}
\title{Check all arguments to px_classification}
\usage{
validate_px_classification_arguments(
  name,
  prestext,
  domain,
  df,
  vs_path,
  agg_paths
)
}
\arguments{
\item{name}{Optional. Name of the classification.}

\item{prestext}{Optional. Presentation text.}

\item{domain}{Optional. Character vector with domain names. Used to link to
PX-file.}

\item{df}{Optional. A data frame with required column 'valuecode' and
optional column 'valuetext', if the codes have texts. Each additional column
represents an aggregation. The column name is the name of the aggregation.
If the column type is character the aggregation levels will be sorted
alphabetically; use factors to control the ordering.}

\item{vs_path}{Optional. Path to a values set (.vs) file.}

\item{agg_paths}{Optional.
\itemize{
\item If NULL, aggregation paths are automatically taken from the
[Aggreg] field in the .vs file.
\item Use a vector of paths to one or more aggregation files (.agg) to
manually choose aggregations.
\item Use character(0) if aggregations from the .vs files should not be
added automatically.
}}
}
\value{
Nothing
}
\description{
Check all arguments to px_classification
}
\keyword{internal}
