% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_px_micro_arguments}
\alias{validate_px_micro_arguments}
\title{Check all arguments to px_micro()}
\usage{
validate_px_micro_arguments(x, out_dir)
}
\arguments{
\item{x}{A px object.}

\item{out_dir}{Directory to save PX-files in.}
}
\value{
Nothing
}
\description{
Check all arguments to px_micro()
}
\keyword{internal}
