% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_px_save_arguments}
\alias{validate_px_save_arguments}
\title{Check all arguments to px_save()}
\usage{
validate_px_save_arguments(x, path, save_data, data_path)
}
\arguments{
\item{x}{A px object.}

\item{path}{Path to file. The file extension determines the format. Can be:
\itemize{
\item \code{.px} to save as a px file
\item \code{.xlsx} to save as an Excel metadata workbook
}}

\item{save_data}{If FALSE, no 'Data' sheet is created in the Excel workbook.
Can only be used if \code{path} is an \code{.xlsx} file.}

\item{data_path}{Path to an \code{.rds} file to save data table at. This is usefull
when saving an Excel workbook where the data has more rows than Excel can
handle. Can only be used if 'path' is an \code{.xlsx} file, and 'save_data' is
TRUE.}
}
\value{
Nothing
}
\description{
Check all arguments to px_save()
}
