% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_pxmake_arguments}
\alias{validate_pxmake_arguments}
\title{Check all pxmake arguments}
\usage{
validate_pxmake_arguments(input, out_path, data, add_totals)
}
\arguments{
\item{input}{Metadata can be provided in one of four ways:
\enumerate{
\item A path to an \code{.xlsx} metadata file.
\item A path to an \code{.rds} file created by \link{metamake}.
\item A named list with two data frames "metadata" and "data" (like
\link{metamake} \code{.rds} file).
\item A data frame with metadata (metadata part of \code{.rds} file by
\link{metamake}.)
}}

\item{out_path}{Path to save output at, either as an \code{.rds} or \code{.px} file.
If NULL, no file is saved.}

\item{data}{Either a data frame, or a path to an \code{.rds} file. If NULL,
the data must be provided as part of the \code{metadata} argument, either in
option 1. as the sheet 'Data' in the Excel metadata workbook, or as the
"data" in option 2 or 3.}

\item{add_totals}{A list of variables to add a 'total' level to. The option
is only available if \code{input} is an \code{.xlsx} file (option 1). The value of the
total level is looked up in 'Variables' xx_elimination. The code for the
level is found in 'Codelists'. The total is a sum of the values in the
variables with \code{pivot = FIGURES} in 'Variables'. NAs are ignored when summing.}
}
\value{
Nothing
}
\description{
Throw an error if there are problems with the function arguments.
}
