% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_pxsave_arguments}
\alias{validate_pxsave_arguments}
\title{Check all arguments to pxsave()}
\usage{
validate_pxsave_arguments(px, out_path)
}
\arguments{
\item{px}{A px object.}
}
\value{
Nothing
}
\description{
Check all arguments to pxsave()
}
