% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_xlsx_metadata}
\alias{validate_xlsx_metadata}
\title{Validate Excel metadata workbook}
\usage{
validate_xlsx_metadata(excel_path)
}
\arguments{
\item{excel_path}{Path to the Excel metadata workbook}
}
\value{
Nothing
}
\description{
Validate Excel metadata workbook
}
