% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_codelists2.R
\name{handle_valuenotes}
\alias{handle_valuenotes}
\alias{valuenote}
\alias{valuenote.px}
\title{Handle VALUENOTE and VALUENOTEX}
\usage{
handle_valuenotes(x, value, type)

valuenote(x, value)

\method{valuenote}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the column 'valuenote' and one or
more of the columns: 'variable-code', 'code', and 'language'. If value is
missing, all current VALUENOTEs are returned. If NULL, all VALUENOTEs are
removed.}

\item{type}{A character string, VALUENOTE or VALUENOTEX}
}
\description{
Inspect or change VALUENOTE.
}
