% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_cells2.R
\name{valuenotex}
\alias{valuenotex}
\alias{valuenotex.px}
\title{VALUENOTEX}
\usage{
valuenotex(x, value)

\method{valuenotex}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the columns 'valuenotex' and one or more of the columns: 'variable-code', 'code', and 'language'. If 'value' is missing, the current VALUENOTEX is returned. If NULL, VALUENOTEX is removed.}
}
\value{
A px object or a data frame
}
\description{
Inspect or change VALUENOTEX.
}
