% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{variable_type}
\alias{variable_type}
\alias{variable_type.px}
\title{VARIABLE-TYPE}
\usage{
variable_type(x, value)

\method{variable_type}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{A data frame with columns 'variable-code' and 'type'. If value
is missing, the current VARIABLE-TYPE is returned. If NULL, all
VARIABLE-TYPE is removed.}
}
\value{
A px object or a data frame.
}
\description{
\code{description_table1("VARIABLE-TYPE")}
}
