% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wrap_varaible_in_list}
\alias{wrap_varaible_in_list}
\title{Convert a variable to a list}
\usage{
wrap_varaible_in_list(df, var)
}
\arguments{
\item{df}{Data frame}

\item{var}{Character. Variable to convert to list}
}
\value{
A data frame
}
\description{
Convert a variable to a list
}
\keyword{internal}
