% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{write_aggregation}
\alias{write_aggregation}
\title{Save classification as .agg file}
\usage{
write_aggregation(aggregation, c, directory)
}
\arguments{
\item{aggregation}{Character, the column name of the aggregation in the
classification}

\item{c}{A classification object}

\item{directory}{Directory to save the file in}
}
\value{
Nothing
}
\description{
Write one aggregation column in a classification to an .agg file.
}
\keyword{internal}
