% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{write_value_set}
\alias{write_value_set}
\title{Save classification as .vs file}
\usage{
write_value_set(c, directory)
}
\arguments{
\item{c}{A classification object}

\item{directory}{Directory to save the file in}
}
\value{
Nothing
}
\description{
Write value set part of classification to a file.
}
\keyword{internal}
