% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{zip_vectors}
\alias{zip_vectors}
\title{Zips list}
\usage{
zip_vectors(v1, v2)
}
\arguments{
\item{v1}{First vector Elements ends up on uneven indexes.}

\item{v2}{Second vector. Elements ends up on even indexes.}
}
\value{
List
}
\description{
Combine two list, by zipping them together in the order \code{v1[1]},
\code{v2[1]}, \code{v1[2]}, \code{v2[2]}, \code{v1[3]}, ...
}
\keyword{internal}
