% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_totals.R
\name{add_total_level_to_var}
\alias{add_total_level_to_var}
\title{Add 'total' level to variable}
\usage{
add_total_level_to_var(df, var, level_name = "Total", sum_var = "value")
}
\arguments{
\item{df}{Data frame}

\item{var}{String, name of variable to total level in}

\item{level_name}{String, total level name}

\item{sum_var}{String, name of variable to sum over}
}
\value{
Data frame
}
\description{
Add a new level to a variables which is the sum of all other levels.
}
