% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_totals.R
\name{add_totals_to_data_table_df}
\alias{add_totals_to_data_table_df}
\title{Add totals to data table}
\usage{
add_totals_to_data_table_df(excel_metadata_path, data_table_df, add_totals)
}
\arguments{
\item{excel_metadata_path}{Path to xlsx workbook with metadata.}

\item{data_table_df}{Data frame with data table in tidy format.}

\item{add_totals}{A list of variables to add a 'total' level to. The option
is only available if \code{input} is an \code{.xlsx} file (option 1). The value of the
total level is looked up in 'Variables' xx_elimination. The code for the
level is found in 'Codelists'. The total is a sum of the values in the
variables with \code{type = FIGURES} in 'Variables'. NAs are ignored when summing.}
}
\description{
Wrapper around \link{add_totals} to get the parameters needed to run the
function from metadata. Add totals can only be run if Excel metadata is used
in \link{pxmake}.
}
