% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_totals.R
\name{add_totals_to_df}
\alias{add_totals_to_df}
\title{Add 'total' levels to multiple variables}
\usage{
add_totals_to_df(df, vars, level_names = "Total", sum_var = "value")
}
\arguments{
\item{df}{Data frame to add total levels to.}

\item{vars}{List of variables to add total levels to.}

\item{level_names}{Names of total levels. Should have length 1 or same length
as \code{vars}.}

\item{sum_var}{String, name of variable to sum over.}
}
\description{
Wrapper around \link{add_total_level_to_var} to add levels to total level to
multiple variables.
}
