% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{change_pivot_variables}
\alias{change_pivot_variables}
\title{Change stub/heading variables}
\usage{
change_pivot_variables(x, pivot, variables)
}
\arguments{
\item{x}{A px object}

\item{pivot}{Pivot type (STUB or HEADING)}

\item{variables}{A character vector of variable codes to change to the pivot
type}
}
\value{
A px object
}
\description{
Change stub/heading variables
}
