% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{charset}
\alias{charset}
\alias{charset.px}
\title{CHARSET}
\usage{
charset(x, value)

\method{charset}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current CHARSET is returned. If NULL, CHARSET is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change CHARSET.
}
