% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{creation_date}
\alias{creation_date}
\alias{creation_date.px}
\title{Set CREATION-DATE}
\usage{
creation_date(x, value)

\method{creation_date}{px}(x, value = format(Sys.time(), "\%Y\%m\%d \%H:\%M"))
}
\description{
Set CREATION-DATE
}
