% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{decimals}
\alias{decimals}
\alias{decimals.px}
\title{DECIMALS}
\usage{
decimals(x, value)

\method{decimals}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current DECIMALS
is returned. If NULL, DECIMALS is set to its default value.}
}
\description{
Inspect or change DECIMALS. DECIMALS cannot be removed because it is a
mandatory keyword.
}
