% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{decimals}
\alias{decimals}
\alias{decimals.px}
\title{Set DECIMALS}
\usage{
decimals(x, value)

\method{decimals}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{A character string}
}
\description{
Set DECIMALS
}
