% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{figures}
\alias{figures}
\alias{figures.px}
\title{FIGURES}
\usage{
figures(x, variable)

\method{figures}{px}(x, variable)
}
\arguments{
\item{x}{A px object}

\item{variable}{Optional. Name of variable to use as FIGRUES. If missing, the
current FIGURES variable is returned.}
}
\value{
A px object or a character string
}
\description{
Inspect or change which variable is used as figures. The previous figures
variable is changed to STUB.
}
\seealso{
\code{\link{stub}} \code{\link{heading}}
}
