% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_px_file.R
\name{format_px_data_as_lines}
\alias{format_px_data_as_lines}
\title{Turn metadata and data cube into text lines that can be written to a px file.}
\usage{
format_px_data_as_lines(metadata, data_cube)
}
\description{
Turn metadata and data cube into text lines that can be written to a px file.
}
