% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_file.R
\name{format_px_object_as_lines}
\alias{format_px_object_as_lines}
\title{Get lines for px-file from px object}
\usage{
format_px_object_as_lines(x)
}
\arguments{
\item{x}{A px object}
}
\value{
A character vector
}
\description{
Get lines for px-file from px object
}
\keyword{internal}
