% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_codelist_metadata}
\alias{get_codelist_metadata}
\title{Get metadata from 'Codelists' sheet in Excel Workbook}
\usage{
get_codelist_metadata(metadata_path)
}
\description{
Get metadata from 'Codelists' sheet in Excel Workbook
}
