% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_data_cube}
\alias{get_data_cube}
\title{Get data cube used in px file format}
\usage{
get_data_cube(metadata_df, data_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}

\item{data_df}{Data frame with data in tidy format.}
}
\value{
A data frame
}
\description{
Get data cube used in px file format
}
