% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_excel_sheet}
\alias{get_excel_sheet}
\title{Get specific sheet from Excel workbook}
\usage{
get_excel_sheet(sheet)
}
\arguments{
\item{sheet}{String. Sheet to read.}

\item{path}{Path to Excel workbook (xls/xlsx).}
}
\value{
A data frame.
}
\description{
Get specific sheet from Excel workbook
}
