% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_figures_variable}
\alias{get_figures_variable}
\title{Get the name of figures variable}
\usage{
get_figures_variable(excel_metadata_path)
}
\arguments{
\item{excel_metadata_path}{Path to Excel workbook with metadata.}
}
\value{
Character
}
\description{
An error is thrown if there is not exactly one figures variable.
}
