% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_df.R
\name{get_main_language}
\alias{get_main_language}
\title{Get the main language from metadata}
\usage{
get_main_language(metadata_df)
}
\arguments{
\item{metadata_df}{Data frame with metadata.}
}
\value{
Character
}
\description{
Get the main language from metadata
}
\keyword{internal}
