% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{get_main_language}
\alias{get_main_language}
\title{Return the main language of the px-file}
\usage{
get_main_language(metadata)
}
\arguments{
\item{metadata}{A data frame with metadata}
}
\description{
Return the main language of the px-file
}
