% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{get_metadata_df}
\alias{get_metadata_df}
\title{Get data frame from metadata argument}
\usage{
get_metadata_df(input, data_df)
}
\arguments{
\item{input}{Metadata can be provided in one of four ways:
\enumerate{
\item A path to an \code{.xlsx} metadata file.
\item A path to an \code{.rds} file created by \link{metamake}.
\item A named list with two data frames "metadata" and "data" (like
\link{metamake} \code{.rds} file).
\item A data frame with metadata (metadata part of \code{.rds} file by
\link{metamake}.)
}}

\item{data_df}{Data frame with data in tidy format.}
}
\description{
Get data frame from metadata argument
}
