% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{get_metadata_template_from_data}
\alias{get_metadata_template_from_data}
\title{Create a minimal metadata template}
\usage{
get_metadata_template_from_data(data_df)
}
\arguments{
\item{data_df}{A data frame with data to create metadata for}
}
\value{
A data frame
}
\description{
Create a minimal metadata template
}
