% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{get_px_metadata_regex}
\alias{get_px_metadata_regex}
\title{Regular expression to parse header in pxfile}
\usage{
get_px_metadata_regex()
}
\value{
Character
}
\description{
Regular expression to parse header in pxfile
}
