% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_variables_metadata}
\alias{get_variables_metadata}
\title{Get metadata about variables}
\usage{
get_variables_metadata(excel_metadata_path)
}
\arguments{
\item{excel_metadata_path}{Path to xlsx workbook with metadata.}
}
\value{
A data frame
}
\description{
Varaibles contains the name the variable in each language, as well as their
type, notes, etc.
}
