% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{guess_file_encoding}
\alias{guess_file_encoding}
\title{Guess encoding of file}
\usage{
guess_file_encoding(path, prefer = list(`UTF-8` = 0.2, `ISO-8859-1` = 0.2))
}
\arguments{
\item{path}{Path to file}

\item{prefer}{Named list of encodings, used to give higher confidence to
specific encodings.}
}
\value{
Character
}
\description{
Guess encoding of file
}
\keyword{internal}
