% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{heading}
\alias{heading}
\alias{heading.px}
\title{Set HEADING}
\usage{
heading(x, variables)

\method{heading}{px}(x, variables)
}
\arguments{
\item{x}{A px object}

\item{variables}{A character vector of variable codes to change to HEADING.}
}
\value{
A px object
}
\description{
Set which variables are used as headings. The order is also changed to the
order of the input variables.
}
