% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{heading}
\alias{heading}
\alias{heading.px}
\title{HEADING}
\usage{
heading(x, variables)

\method{heading}{px}(x, variables)
}
\arguments{
\item{x}{A px object}

\item{variables}{Optional. A character vector of variable names to change to
STUB. This also changes the HEADING order. With names in \code{variables}
becoming 1, 2, ... If missing, the current HEADING variables are returned.}
}
\value{
A px object or a character vector
}
\description{
Inspect or change HEADING.
}
\seealso{
\code{\link{stub}} \code{\link{figures}}
}
