% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{last_updated}
\alias{last_updated}
\alias{last_updated.px}
\title{LAST-UPDATED}
\usage{
last_updated(x, value)

\method{last_updated}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current
LAST-UPDATED is returned. If NULL, LAST-UPDATED is removed.}
}
\description{
Inspect or change LAST-UPDATED.
}
