% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template.R
\name{make_template}
\alias{make_template}
\title{Make template}
\usage{
make_template(
  data_df,
  languages = c("en"),
  time_variable = NULL,
  figures_variable = NULL,
  out_path = NULL
)
}
\arguments{
\item{data_df}{A data frame to create metadata for}

\item{languages}{A vector of languagecodes. The first value is the main
language.}

\item{time_variable}{String. Variable to use as time variable}

\item{figures_variable}{String. Variable to use as figures variable}

\item{out_path}{Path to save .xlsx file at. If not specified a temporary file
is created.}
}
\value{
A data frame
}
\description{
Create a minimal metadata template for a data frame.
\code{make_template} will be deprecated in the future when a px object has been
implemented (#179).
}
