% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{matrix}
\alias{matrix}
\alias{matrix.px}
\title{Set MATRIX}
\usage{
matrix(x, value)

\method{matrix}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{A character string}
}
\description{
Set MATRIX
}
