% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{matrix}
\alias{matrix}
\alias{matrix.px}
\title{MATRIX}
\usage{
matrix(x, value)

\method{matrix}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current MATRIX is
returned. If NULL, MATRIX is removed.}
}
\description{
Inspect or change MATRIX.
}
