% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{metamake}
\alias{metamake}
\title{Create an Excel metadata workbook from a px-file}
\usage{
metamake(px_file_path, out_path)
}
\arguments{
\item{px_file_path}{Path to px file}

\item{out_path}{Path to save file at}
}
\value{
Nothing
}
\description{
Turn a px-file into an Excel metadata workbook. If pxmake() is run on that
workbook it turns back into the same px-file.
}
