% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{metamake}
\alias{metamake}
\title{Create an Excel metadata workbook from a px-file}
\usage{
metamake(input, out_path, data_table_path = NULL)
}
\arguments{
\item{input}{Input can be provided in one of three ways:
\enumerate{
\item A path to a \code{.px} file.
\item A path to a \code{.rds} file created by \link{pxmake}.
\item A named list with two data frames "metadata" and "data_table" (same as
option 2).
}}

\item{out_path}{Path to save metadata at. Use \code{.xlsx} extension to save
as an Excel workbook. Use \code{.rds} to save as an rds file.}

\item{data_table_path}{Path to save data table as an .rds file. If NULL, the
data table is saved as the sheet 'Data' in the Excel metadata workbook.}
}
\value{
Nothing
}
\description{
Create an Excel metadata workbook from a px-file
}
\seealso{
\link{pxmake}
}
