% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{metamake}
\alias{metamake}
\title{Create a pxobject from a px-file}
\usage{
metamake(input, out_path = NULL, data_path = NULL, create_data = TRUE)
}
\arguments{
\item{input}{Input can be provided in one of three ways:
\enumerate{
\item A path to a \code{.px} file.
\item A path to a \code{.rds} file created by \link{pxmake}.
}}

\item{out_path}{Path to save metadata at. Use \code{.xlsx} extension to save
as an Excel workbook. Use \code{.rds} to save as an rds file. If NULL, no file is
saved.}

\item{data_path}{Path to save data as an .rds file. If NULL, the data is
saved as the sheet 'Data' in the Excel metadata workbook.}

\item{create_data}{Logic. If FALSE the data table is not generated. This can
be used to only generate metadata.}
}
\value{
Returns rds object invisibly.
}
\description{
Create a pxobject from a px-file
}
\seealso{
\link{pxmake}
}
