% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamake.R
\name{metamake}
\alias{metamake}
\title{Create an Excel metadata workbook from a px-file}
\usage{
metamake(px_path, xlsx_path, rds_data_path = NULL, overwrite_xlsx = TRUE)
}
\arguments{
\item{px_path}{Path to a px file}

\item{xlsx_path}{Path to save xlsx file at}

\item{rds_data_path}{Path to save data cube as rds file. If NULL the data
cube is added in the 'Data' sheet in the Excel metadata workbook.}

\item{overwrite_xlsx}{Should existing metadata workbook be overwritten?}
}
\value{
Nothing
}
\description{
Turn a px-file into an Excel metadata workbook. If pxmake() is run on that
workbook it turns back into an equivalent px-file.
}
