% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromake.R
\name{micromake}
\alias{micromake}
\title{Create microdata}
\usage{
micromake(data_df, metadata_path, out_dir = temp_dir())
}
\arguments{
\item{data_df}{A data frame.}

\item{metadata_path}{An Excel workbook created by metamake.}

\item{out_dir}{Directory to save px-files in.}
}
\value{
Nothing
}
\description{
Turn a dataset and its metadata into a series of px-files, one for each
variable in the original dataset, except time vars.
}
