% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromake.R
\name{micromake}
\alias{micromake}
\title{Create microdata}
\usage{
micromake(px, out_dir = NULL)
}
\arguments{
\item{px}{A px object.}

\item{out_dir}{Directory to save px files in.}
}
\value{
Nothing
}
\description{
Turn a px object into many px files, one for each variable except time vars.
}
