% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromake.R
\name{micromake}
\alias{micromake}
\title{Create microdata}
\usage{
micromake(input, out_dir)
}
\arguments{
\item{input}{Input can be provided in one of four ways:
\enumerate{
\item A path to a \code{.px} file.
\item A path to a \code{.rds} file created by \link{pxmake}.
\item A named list with two data frames "metadata" and "data" (same as option 2).
\item A data frame with data. A minimal metadata template will be created.
}}

\item{out_dir}{Directory to save px-files in.}
}
\value{
Nothing
}
\description{
Turn a dataset and its metadata into a series of px-files, one for each
variable in the original dataset, except time vars.
}
