% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromake.R
\name{micromake}
\alias{micromake}
\title{Create micro px files}
\usage{
micromake(x, out_dir = NULL, keyword_values = NULL)
}
\arguments{
\item{x}{A px object.}

\item{out_dir}{Directory to save px files in.}

\item{keyword_values}{Optional. A data frame with column 'variable' and one
or more of: 'contents', 'title', 'description', and 'matrix'. The columns
with be added as keywords to the table for each non-HEADING varabe that
match the 'variable' column. It probably work for other keywords as well.
Currently doesn't support multiple languages.}
}
\value{
Nothing
}
\description{
Split one px object into many small px files (micro files), with count of
the variables in it.
}
\details{
The HEADING variables are use in all the micro files, and a file is created
for each non-HEADING variable. The new px files are saved in a directory
specified by \code{out_dir}.
}
