% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromake.R
\name{micromake2}
\alias{micromake2}
\title{Title}
\usage{
micromake2(data_df, metadata_path, out_dir = temp_dir())
}
\arguments{
\item{data_df}{A data frame.}

\item{out_dir}{Directory to save px-files in.}

\item{metadata}{An Excel workbook created by metamake.}
}
\value{
Nothing
}
\description{
Description
}
