% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_codelists1.R
\name{handle_codelists}
\alias{handle_codelists}
\alias{precision}
\alias{precision.px}
\title{Handle codelists changes}
\usage{
handle_codelists(x, value, number, column)

precision(x, value)

\method{precision}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the column 'precision' and one or
more of the columns: 'variable-code' and 'code'. If value is missing, all
current precisions are returned. If NULL, all precisions are removed.}

\item{number}{The number of the codelist (1 or 2)}

\item{column}{The name of the codelists column}
}
\description{
Inspect or change PRECISION.
}
