% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{px}
\alias{px}
\title{Create a px object}
\usage{
px(input, data = NULL)
}
\arguments{
\item{input}{Path to px-file, path to an Excel metadata workbook.}

\item{data}{Either a data frame or a path to an \code{.rds} file with a data frame.
This can only be used if the \code{input} argument is an Excel metadata workbook.
If NULL, the data should be provided in the 'Data' sheet of the Excel
workbook.}
}
\value{
A px object invisibly.
}
\description{
Create a px object from a px-file, an Excel metadata workbook, or a list with
a specific structure.
}
