% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{px}
\alias{px}
\title{Create a px object}
\usage{
px(input, data = NULL)
}
\arguments{
\item{input}{Path to px file, path to an Excel metadata workbook, a data
frame or path to an .rds file with a data frame. If input is a data frame, a
px object with minimal metadata is created.}

\item{data}{Either a data frame or a path to an \code{.rds} file with a data frame.
This can only be used if \code{input} is an Excel metadata workbook.}
}
\value{
A px object
}
\description{
Create a px object from a px file, an Excel metadata workbook, or a data
frame.
}
