% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_aggregallowed}
\alias{px_aggregallowed}
\alias{px_aggregallowed.px}
\title{AGGREGALLOWED}
\usage{
px_aggregallowed(x, value)

\method{px_aggregallowed}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current AGGREGALLOWED is returned. If NULL, AGGREGALLOWED is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change AGGREGALLOWED.
}
\examples{
# Set AGGREGALLOWED
x1 <-
   px(population_gl) |>
   px_aggregallowed('NO')

# Print AGGREGALLOWED
px_aggregallowed(x1)

# Remove AGGREGALLOWED
x2 <- px_aggregallowed(x1, NULL)
px_aggregallowed(x2)
}
