% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_aggregallowed}
\alias{px_aggregallowed}
\alias{px_aggregallowed.px}
\title{AGGREGALLOWED}
\usage{
px_aggregallowed(x, value)

\method{px_aggregallowed}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current AGGREGALLOWED is returned. If NULL, AGGREGALLOWED is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change AGGREGALLOWED.
}
\examples{
# Set AGGREGALLOWED
x1 <-
   px(population_gl) |>
   px_aggregallowed('NO')

# Print AGGREGALLOWED
px_aggregallowed(x1)

# Remove AGGREGALLOWED
x2 <- px_aggregallowed(x1, NULL)
px_aggregallowed(x2)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A37\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C601\%2C0\%5D}{Statistics Sweden's documentation}
}
