% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_axis_version}
\alias{px_axis_version}
\alias{px_axis_version.px}
\title{AXIS-VERSION}
\usage{
px_axis_version(x, value)

\method{px_axis_version}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current AXIS-VERSION is returned. If NULL, AXIS-VERSION is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change AXIS-VERSION.
}
