% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_baseperiod}
\alias{px_baseperiod}
\alias{px_baseperiod.px}
\title{BASEPERIOD}
\usage{
px_baseperiod(x, value)

\method{px_baseperiod}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current BASEPERIOD is returned. If NULL, BASEPERIOD is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change BASEPERIOD.
}
\examples{
# Set BASEPERIOD for all languages
x1 <-
  px(population_gl) |>
  px_baseperiod('year')

# Print BASEPERIOD
px_baseperiod(x1)

# Set BASEPERIOD for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_baseperiod(tribble(~language, ~value,
                      'en', 'year',
                      'kl', 'ukioq'))
px_baseperiod(x2)

# Remove BASEPERIOD
x3 <- px_baseperiod(x2, NULL)
px_baseperiod(x3)
}
