% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_acrosscells.R
\name{px_cellnote}
\alias{px_cellnote}
\alias{px_cellnote.px}
\title{CELLNOTE}
\usage{
px_cellnote(x, value)

\method{px_cellnote}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with columns 'cellnote' and one or more columns with the names of the STUB and HEADING variables. The 'cellnote' column is the cellnote text, and the STUB/HEADING columns control which cells the note applies to. Use '*' if the note applies to all cells for a variable.If 'value' is missing, the current CELLNOTE is returned. If NULL, CELLNOTE is removed.}
}
\value{
A px object, or a data frame.
}
\description{
Inspect or change CELLNOTE.
}
