% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_cfprices}
\alias{px_cfprices}
\alias{px_cfprices.px}
\title{CFPRICES}
\usage{
px_cfprices(x, value)

\method{px_cfprices}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current CFPRICES is returned. If NULL, CFPRICES is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change CFPRICES.
}
\examples{
# Set CFPRICES for all languages
x1 <-
  px(population_gl) |>
  px_cfprices('C')

# Print CFPRICES
px_cfprices(x1)

# Set CFPRICES for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_cfprices(tribble(~language, ~value,
                      'en', 'C',
                      'kl', 'F'))
px_cfprices(x2)

# Remove CFPRICES
x3 <- px_cfprices(x2, NULL)
px_cfprices(x3)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A40\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C178\%2C0\%5D}{Statistics Sweden's documentation}
}
