% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_charset}
\alias{px_charset}
\alias{px_charset.px}
\title{CHARSET}
\usage{
px_charset(x, value)

\method{px_charset}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current CHARSET is returned. If NULL, CHARSET is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change CHARSET.
}
\examples{
# Set CHARSET
x1 <-
   px(population_gl) |>
   px_charset('ANSI')

# Print CHARSET
px_charset(x1)

# Remove CHARSET
x2 <- px_charset(x1, NULL)
px_charset(x2)
}
