% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_contents}
\alias{px_contents}
\alias{px_contents.px}
\title{CONTENTS}
\usage{
px_contents(x, value)

\method{px_contents}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current CONTENTS is returned. If NULL, an error is thrown because CONTENTS cannot be removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change CONTENTS.
}
\examples{
# Set CONTENTS for all languages
x1 <-
  px(population_gl) |>
  px_contents('Population')

# Print CONTENTS
px_contents(x1)

# Set CONTENTS for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_contents(tribble(~language, ~value,
                      'en', 'Population',
                      'kl', 'Innuttaasut'))
px_contents(x2)
}
