% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_copyright}
\alias{px_copyright}
\alias{px_copyright.px}
\title{COPYRIGHT}
\usage{
px_copyright(x, value)

\method{px_copyright}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current COPYRIGHT is returned. If NULL, COPYRIGHT is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change COPYRIGHT.
}
\examples{
# Set COPYRIGHT
x1 <-
   px(population_gl) |>
   px_copyright('YES')

# Print COPYRIGHT
px_copyright(x1)

# Remove COPYRIGHT
x2 <- px_copyright(x1, NULL)
px_copyright(x2)
}
