% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_copyright}
\alias{px_copyright}
\alias{px_copyright.px}
\title{COPYRIGHT}
\usage{
px_copyright(x, value)

\method{px_copyright}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current COPYRIGHT is returned. If NULL, COPYRIGHT is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change COPYRIGHT.
}
\examples{
# Set COPYRIGHT
x1 <-
   px(population_gl) |>
   px_copyright('YES')

# Print COPYRIGHT
px_copyright(x1)

# Remove COPYRIGHT
x2 <- px_copyright(x1, NULL)
px_copyright(x2)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A42\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C302\%2C0\%5D}{Statistics Sweden's documentation}
}
