% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{px_data}
\alias{px_data}
\alias{px_data.px}
\title{DATA}
\usage{
px_data(x, value)

\method{px_data}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame. If missing, the current DATA is returned.
If NULL, all data rows are removed.}
}
\value{
A px object or a data frame
}
\description{
Inspect or change DATA.
}
