% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{px_data}
\alias{px_data}
\alias{px_data.px}
\title{DATA}
\usage{
px_data(x, value)

\method{px_data}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame. If missing, the current DATA is returned.
If NULL, all data rows are removed.}
}
\value{
A px object or data frame.
}
\description{
Inspect or change DATA.
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A32\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C502\%2C0\%5D}{Statistics Sweden's documentation}
}
