% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_decimals}
\alias{px_decimals}
\alias{px_decimals.px}
\title{DECIMALS}
\usage{
px_decimals(x, value)

\method{px_decimals}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current DECIMALS is returned. If NULL, an error is thrown because DECIMALS cannot be removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change DECIMALS.
}
\examples{
# Set DECIMALS
x1 <-
   px(population_gl) |>
   px_decimals('3')

# Print DECIMALS
px_decimals(x1)
}
