% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_description}
\alias{px_description}
\alias{px_description.px}
\title{DESCRIPTION}
\usage{
px_description(x, value)

\method{px_description}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current DESCRIPTION is returned. If NULL, DESCRIPTION is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change DESCRIPTION.
}
\examples{
# Set DESCRIPTION for all languages
x1 <-
  px(population_gl) |>
  px_description('Population')

# Print DESCRIPTION
px_description(x1)

# Set DESCRIPTION for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_description(tribble(~language, ~value,
                      'en', 'Population',
                      'kl', 'Innuttaasut'))
px_description(x2)

# Remove DESCRIPTION
x3 <- px_description(x2, NULL)
px_description(x3)
}
\seealso{
\code{\link[=px_title]{px_title()}}
}
