% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_descriptiondefault}
\alias{px_descriptiondefault}
\alias{px_descriptiondefault.px}
\title{DESCRIPTIONDEFAULT}
\usage{
px_descriptiondefault(x, value)

\method{px_descriptiondefault}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current DESCRIPTIONDEFAULT is returned. If NULL, DESCRIPTIONDEFAULT is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change DESCRIPTIONDEFAULT.
}
\examples{
# Set DESCRIPTIONDEFAULT
x1 <-
   px(population_gl) |>
   px_descriptiondefault('YES')

# Print DESCRIPTIONDEFAULT
px_descriptiondefault(x1)

# Remove DESCRIPTIONDEFAULT
x2 <- px_descriptiondefault(x1, NULL)
px_descriptiondefault(x2)
}
