% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable2.R
\name{px_elimination}
\alias{px_elimination}
\alias{px_elimination.px}
\title{ELIMINATION}
\usage{
px_elimination(x, value)

\method{px_elimination}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string or data frame.
\itemize{
\item Use character to set ELIMINATION for all languages and variables.
\item Use data frame with columns 'variable-code', 'language' and
'elimination' to set ELIMINATION for specific variables.
\item If missing, the current ELIMINATION is returned.
\item If NULL, ELIMINATION is removed for all variables.
}}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change ELIMINATION.
}
\examples{
# Set ELIMINATION for all languages
x1 <-
  px(population_gl) |>
  px_elimination('YES')

# Print ELIMINATION
px_elimination(x1)

# Set ELIMINATION for individual variables
library(tibble)
x2 <-
  x1 |>
  px_elimination(tribble(~`variable-code`, ~elimination,
                    'gender', 'All',
                    'age',    'Total'))
px_elimination(x2)

# Set ELIMINATION for individual languages
x3 <-
  x2 \%>\%
  px_languages(c('en', 'kl')) |>
  px_elimination(tribble(~`variable-code`, ~language, ~elimination,
                    'gender',    'en',      'All_en',
                    'gender',    'kl',      'All_kl',
                    'age',       'en',      'Total_en'))
px_elimination(x3)

# Remove ELIMINATION
x4 <- px_elimination(x3, NULL)
px_elimination(x4)
}
