% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{px_from_excel}
\alias{px_from_excel}
\title{Turn Excel workbook into a px object}
\usage{
px_from_excel(excel_path, data = NULL)
}
\arguments{
\item{excel_path}{Path to Excel metadata workbook.}
}
\value{
A px object.
}
\description{
Turn Excel workbook into a px object
}
