% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{px_heading}
\alias{px_heading}
\alias{px_heading.px}
\title{HEADING}
\usage{
px_heading(x, value)

\method{px_heading}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character vector of variable names to change to
STUB. This also changes the HEADING order. With names in \code{variables}
becoming 1, 2, ... If missing, the current HEADING variables are returned.}
}
\value{
A px object or a character vector.
}
\description{
Inspect or change HEADING.
}
\examples{
x1 <- px(population_gl)

# Print HEADING
px_heading(x1)

# Add 'gender' to HEADING
x2 <- px_heading(x1, 'gender')
px_heading(x2)

# Change order of HEADING
x3 <- px_heading(x2, 'year')
px_heading(x3)
}
\seealso{
\code{\link{px_stub}} \code{\link{px_figures}}
}
