% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_infofile}
\alias{px_infofile}
\alias{px_infofile.px}
\title{INFOFILE}
\usage{
px_infofile(x, value)

\method{px_infofile}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current INFOFILE is returned. If NULL, INFOFILE is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change INFOFILE.
}
\examples{
# Set INFOFILE for all languages
x1 <-
  px(population_gl) |>
  px_infofile('infofile_en')

# Print INFOFILE
px_infofile(x1)

# Set INFOFILE for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_infofile(tribble(~language, ~value,
                      'en', 'infofile_en',
                      'kl', 'infofile_kl'))
px_infofile(x2)

# Remove INFOFILE
x3 <- px_infofile(x2, NULL)
px_infofile(x3)
}
