% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_languages.R
\name{px_languages}
\alias{px_languages}
\alias{px_languages.px}
\title{LANGUAGES}
\usage{
px_languages(x, value)

\method{px_languages}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character vector. If missing, the current LANGUAGES
are returned. If NULL, LANGUAGES are removed.}
}
\description{
Inspect or change LANGUAGES.
}
\details{
If LANGUAGE is defined it should be one of the values in LANGUAGES.
}
\seealso{
\code{\link{language}}
}
