% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_last_updated}
\alias{px_last_updated}
\alias{px_last_updated.px}
\title{LAST-UPDATED}
\usage{
px_last_updated(x, value)

\method{px_last_updated}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current LAST-UPDATED is returned. If NULL, LAST-UPDATED is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change LAST-UPDATED.
}
