% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_link}
\alias{px_link}
\alias{px_link.px}
\title{LINK}
\usage{
px_link(x, value)

\method{px_link}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current LINK is returned. If NULL, LINK is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change LINK.
}
\examples{
# Set LINK for all languages
x1 <-
  px(population_gl) |>
  px_link('https://stat.gl/?lang=en')

# Print LINK
px_link(x1)

# Set LINK for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_link(tribble(~language, ~value,
                      'en', 'https://stat.gl/?lang=en',
                      'kl', 'https://stat.gl/'))
px_link(x2)

# Remove LINK
x3 <- px_link(x2, NULL)
px_link(x3)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A49\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C453\%2C0\%5D}{Statistics Sweden's documentation}
}
