% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_cells1.R
\name{px_order}
\alias{px_order}
\alias{px_order.px}
\title{ORDER}
\usage{
px_order(x, value)

\method{px_order}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the columns 'px_order' and one or more of the columns: 'variable-code', and 'code'. If 'value' is missing, the current PX_ORDER is returned. If NULL, PX_ORDER is removed.}
}
\value{
A px object or a data frame
}
\description{
Inspect or change PX_ORDER.
}
