% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_showdecimals}
\alias{px_showdecimals}
\alias{px_showdecimals.px}
\title{SHOWDECIMALS}
\usage{
px_showdecimals(x, value)

\method{px_showdecimals}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current SHOWDECIMALS is returned. If NULL, SHOWDECIMALS is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change SHOWDECIMALS.
}
