% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_tableid}
\alias{px_tableid}
\alias{px_tableid.px}
\title{TABLEID}
\usage{
px_tableid(x, value)

\method{px_tableid}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current TABLEID is returned. If NULL, TABLEID is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change TABLEID.
}
\examples{
# Set TABLEID
x1 <-
   px(population_gl) |>
   px_tableid('POPGL')

# Print TABLEID
px_tableid(x1)

# Remove TABLEID
x2 <- px_tableid(x1, NULL)
px_tableid(x2)
}
