% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{px_timeval}
\alias{px_timeval}
\alias{px_timeval.px}
\title{TIMEVAL}
\usage{
px_timeval(x, value)

\method{px_timeval}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current TIMEVAL is returned. If NULL, TIMEVAL is removed.}
}
\value{
A px object or a character string.
}
\description{
Inspect or change TIMEVAL.

There can only be one time variable.
}
\examples{
# Set TIMEVAL
x1 <-
   px(population_gl) |>
   px_timeval('year')

# Print TIMEVAL
px_timeval(x1)

# Remove TIMEVAL
x2 <- px_timeval(x1, NULL)
px_timeval(x2)
}
\seealso{
\href{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf#\%5B\%7B\%22num\%22\%3A54\%2C\%22gen\%22\%3A0\%7D\%2C\%7B\%22name\%22\%3A\%22XYZ\%22\%7D\%2C184\%2C181\%2C0\%5D}{Statistics Sweden's documentation}
}
