% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_title}
\alias{px_title}
\alias{px_title.px}
\title{TITLE}
\usage{
px_title(x, value)

\method{px_title}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current TITLE is returned. If NULL, TITLE is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change TITLE.
}
