% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_title}
\alias{px_title}
\alias{px_title.px}
\title{TITLE}
\usage{
px_title(x, value)

\method{px_title}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current TITLE is returned. If NULL, TITLE is removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change TITLE.

TITLE can only be removed if DESCRIPTION is set.
}
\examples{
# Set TITLE for all languages
x1 <-
  px(population_gl) |>
  px_title('Population GR')

# Print TITLE
px_title(x1)

# Set TITLE for individual languages
library(tibble)
x2 <-
  x1 |>
  px_languages(c('en', 'kl')) |>
  px_title(tribble(~language, ~value,
                      'en', 'Population GR',
                      'kl', 'Innuttaasut KL'))
px_title(x2)
}
\seealso{
\code{\link[=px_description]{px_description()}}
}
