% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table2.R
\name{px_units}
\alias{px_units}
\alias{px_units.px}
\title{UNITS}
\usage{
px_units(x, value)

\method{px_units}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string to set the value for all languages or a data frame with columns 'language' and 'value' to set it for specific languages. If 'value' is missing, the current UNITS is returned. If NULL, an error is thrown because UNITS cannot be removed.}
}
\value{
A px object, a character string, or a data frame.
}
\description{
Inspect or change UNITS. UNITS cannot be removed because it is a mandatory keyword.
}
