% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table1.R
\name{px_update_frequency}
\alias{px_update_frequency}
\alias{px_update_frequency.px}
\title{UPDATE-FREQUENCY}
\usage{
px_update_frequency(x, value)

\method{px_update_frequency}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A character string. If missing, the current UPDATE-FREQUENCY is returned. If NULL, UPDATE-FREQUENCY is removed.}
}
\value{
A px object or a character string
}
\description{
Inspect or change UPDATE-FREQUENCY.
}
