% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_cells2.R
\name{px_valuenote}
\alias{px_valuenote}
\alias{px_valuenote.px}
\title{VALUENOTE}
\usage{
px_valuenote(x, value)

\method{px_valuenote}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the columns 'valuenote' and one or more of the columns: 'variable-code', 'code', and 'language'. If 'value' is missing, the current VALUENOTE is returned. If NULL, VALUENOTE is removed.}
}
\value{
A px object or a data frame
}
\description{
Inspect or change VALUENOTE.
}
