% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_cells2.R
\name{px_values}
\alias{px_values}
\alias{px_values.px}
\title{VALUES}
\usage{
px_values(x, value)

\method{px_values}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the columns 'values' and one or more of the columns: 'variable-code', 'code', and 'language'. If 'value' is missing, the current VALUES is returned. If NULL, VALUES is removed.}
}
\value{
A px object or a data frame
}
\description{
Inspect or change VALUES.
VALUES cannot be removed because it is a mandatory keyword.
}
