% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake-package.R
\docType{package}
\name{pxmake-package}
\alias{pxmake}
\alias{pxmake-package}
\alias{pxmake-package.R}
\title{pxmake: Create px files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create and modify px files in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StatisticsGreenland/pxmake}
  \item \url{https://statisticsgreenland.github.io/pxmake/}
}

}
\author{
\strong{Maintainer}: Lars Pedersen \email{larp@stat.gl}

Authors:
\itemize{
  \item Johan Ejstrud \email{johan@ejstrud.com}
}

Other contributors:
\itemize{
  \item Statistics Greenland (https://stat.gl/) [copyright holder]
}

}
\keyword{internal}
