% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake-package.R
\docType{package}
\name{pxmake-package}
\alias{pxmake}
\alias{pxmake-package}
\title{pxmake: Create px files}
\description{
Create px files by using a data set and metadata from and excel sheet.
}
\author{
\strong{Maintainer}: Lars Pedersen \email{larp@stat.gl}

Authors:
\itemize{
  \item Johan Ejstrud \email{johan@ejstrud.com}
}

}
\keyword{internal}
