% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake-package.R
\docType{package}
\name{pxmake-package}
\alias{pxmake}
\alias{pxmake-package}
\alias{pxmake-package.R}
\title{pxmake: Make PX-Files in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create PX-files from scratch or read and modify existing ones. Includes a function for every PX keyword, making metadata manipulation simple and human-readable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StatisticsGreenland/pxmake}
  \item \url{https://statisticsgreenland.github.io/pxmake/}
  \item Report bugs at \url{https://github.com/StatisticsGreenland/pxmake/issues}
}

}
\author{
\strong{Maintainer}: Johan Ejstrud \email{johan@ejstrud.com}

Authors:
\itemize{
  \item Lars Pedersen \email{larp@stat.gl}
}

Other contributors:
\itemize{
  \item Statistics Greenland (https://stat.gl/) [copyright holder]
}

}
\keyword{internal}
