% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{pxmake}
\alias{pxmake}
\title{Create pxfile}
\usage{
pxmake(source_data_path, metadata_path, pxfile_path)
}
\arguments{
\item{source_data_path}{Path to \code{.rds} file with data source.}

\item{metadata_path}{Path to Excel workbook with metadata.}

\item{pxfile_path}{Path to save px file at.}
}
\value{
Nothing.
}
\description{
\code{pxmake()} creates a px file by combine source data from a \code{.rds} file and
metadata from an Excel workbook.
}
