% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxmake.R
\name{pxmake}
\alias{pxmake}
\title{Create a px-file from metadata and data}
\usage{
pxmake(input, out_path = NULL, data = NULL, add_totals = NULL)
}
\arguments{
\item{input}{Path to a \code{.xlsx} metadata file.}

\item{out_path}{Path to save output at, either as an \code{.rds}, \code{.px} or \code{.xlsx}
file. If NULL, no file is saved.}

\item{data}{Either a data frame, or a path to an \code{.rds} file. If NULL,
the data must be provided as part of the \code{metadata} argument, either in
option 1. as the sheet 'Data' in the Excel metadata workbook, or as the
"data" in option 2.}

\item{add_totals}{A list of variables to add a 'total' level to. The option
is only available if \code{input} is an \code{.xlsx} file (option 1). The value of the
total level is looked up in 'Variables' xx_elimination. The code for the
level is found in 'cells'. The total is a sum of the values in the
variables with \code{pivot = FIGURES} in 'Variables'. NAs are ignored when summing.}
}
\value{
Returns px object invisibly.
}
\description{
This function will be deprecated in the future. Use \code{px} and \code{pxsave} instead.
}
\seealso{
\link{metamake}
}
