% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_px_file}
\alias{read_px_file}
\title{Get px file content as lines}
\usage{
read_px_file(px_path)
}
\arguments{
\item{px_path}{Path to a px file}
}
\value{
A character vector
}
\description{
Read file with correct encoding.
}
\keyword{internal}
