% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{save_px_as_r_script}
\alias{save_px_as_r_script}
\title{Save px object as an R script}
\usage{
save_px_as_r_script(x, path)
}
\arguments{
\item{x}{A px object}

\item{path}{Path to save R script at}
}
\value{
Nothing
}
\description{
Running the R script creates the px object.
}
\keyword{internal}
