% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{save_px_as_xlsx}
\alias{save_px_as_xlsx}
\title{Save px object as an Excel workbook}
\usage{
save_px_as_xlsx(x, path, save_data, data_path)
}
\arguments{
\item{x}{A px object}

\item{path}{Path to save Excel workbook}

\item{save_data}{If FALSE, no 'Data' sheet is created in the Excel workbook.
Can only be used if \code{path} is an \code{.xlsx} file.}

\item{data_path}{Path to an \code{.rds} or \code{.parquet} file to save data table at.
This is usefull when saving an Excel workbook where the data has more rows
than Excel can handle. Can only be used if 'path' is an \code{.xlsx} file, and
'save_data' is TRUE.}
}
\value{
Nothing
}
\description{
Save px object as an Excel workbook
}
\keyword{internal}
