% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{stub}
\alias{stub}
\alias{stub.px}
\title{STUB}
\usage{
stub(x, variables)

\method{stub}{px}(x, variables)
}
\arguments{
\item{x}{A px object}

\item{variables}{Optional. A character vector of variable names to change to
STUB. If missing, the current STUB variables are returned.}
}
\value{
A px object or a character vector
}
\description{
Inspect or change which variables are used as stubs. The stub order is also
changed to the order of the input variables.
}
\seealso{
\code{\link{heading}} \code{\link{figures}}
}
