% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{stub}
\alias{stub}
\alias{stub.px}
\title{Set STUB}
\usage{
stub(x, variables)

\method{stub}{px}(x, variables)
}
\arguments{
\item{x}{A px object}

\item{variables}{A character vector of variable codes to change to STUB.}
}
\value{
A px object
}
\description{
Set which variables are used as stubs. The order is also changed to the order
of the input variables.
}
