% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variable1.R
\name{timeval}
\alias{timeval}
\alias{timeval.px}
\title{TIMEVAL}
\usage{
timeval(x, variable)

\method{timeval}{px}(x, variable)
}
\arguments{
\item{x}{A px object}

\item{variable}{Optional. Name of variable to use as TIME. If missing, the
current TIME variable is returned.}
}
\value{
A px object or a character string
}
\description{
Inspect or change which variable is used as timeval. There can only be one
time variable.
}
