% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_px_arguments}
\alias{validate_px_arguments}
\title{Check all arguments to px()}
\usage{
validate_px_arguments(input, data)
}
\arguments{
\item{input}{Path to px file, path to an Excel metadata workbook, a data
frame or path to an \code{.rds} or \code{.parquet} file with a data frame. If input is
a data frame or NULL, a px object with minimal metadata is created.}

\item{data}{Either a data frame or a path to an \code{.rds} or \code{.parquet} file
with a data frame. This can only be used if \code{input} is an Excel metadata
workbook.}
}
\value{
Nothing
}
\description{
Check all arguments to px()
}
\keyword{internal}
