% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_warning_error.R
\name{validate_pxsave_arguments}
\alias{validate_pxsave_arguments}
\title{Check all arguments to pxsave()}
\usage{
validate_pxsave_arguments(x, path)
}
\arguments{
\item{x}{A px object.}

\item{path}{Path to file. The file extension determines the format. Can be:
\itemize{
\item \code{.px} to save as a px file
\item \code{.xlsx} to save as an Excel metadata workbook
}}
}
\value{
Nothing
}
\description{
Check all arguments to pxsave()
}
