% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_codelists2.R
\name{valuenote}
\alias{valuenote}
\alias{valuenote.px}
\title{VALUENOTE}
\usage{
valuenote(x, value)

\method{valuenote}{px}(x, value)
}
\arguments{
\item{x}{A px object}

\item{value}{Optional. A data frame with the column 'valuenote' and one or
more of the columns: 'variable-code', 'code', and 'language'. If value is
missing, all current VALUENOTEs are returned. If NULL, all VALUENOTEs are
removed.}
}
\description{
Inspect or change VALUENOTE.
}
